Scoped Helper Verbs

Scoped helper verbs included in this R Documentation file allow for targeted commands on specified columns. They also rename the ensuing output to conform to my preferred style. The commands here are multiple and explained in the details section below.

diff_at(data, x, l = 1)

r1sd_at(data, x, na = TRUE)

r2sd_at(data, x, na = TRUE)

Arguments

data

a data frame

x

a vector, likely in your data frame

l

The order of lags for calculating differences in diff_at. Applicable only to diff_at.

na

what to do with NAs in the vector. Defaults to TRUE (i.e. passes over the missing observations). Not applicable to diff_at.

Value

The function returns a numeric vector rescaled with a mean of 0 and a standard deviation of .5.

Details

r1sd_at is a wrapper for mutate_at and rename_at from dplyr. It both rescales the supplied vectors to new vectors and renames the vectors to each have a prefix of s_. Note the rescaling here is just by one standard deviation and not two.

r2sd_at is a wrapper for mutate_at and rename_at from dplyr. It both rescales the supplied vectors to new vectors and renames the vectors to each have a prefix of z_. Note the rescaling here is by two standard deviations and not one.

By default, na.rm is set to TRUE. If you have missing data, the function will just pass over them.

Gelman (2008) argues that rescaling by two standard deviations puts regression inputs on roughly the same scale no matter their original scale. This allows for some honest, if preliminary, assessment of relative effect sizes from the regression output. This does that, but without requiring the rescale function from arm. I'm trying to reduce the packages on which my workflow relies.

Importantly, I tend to rescale only the ordinal and interval inputs and leave the binary inputs as 0/1. So, my r2sd function doesn't have any of the fancier if-else statements that Gelman's rescale function has.

References

Gelman, Andrew. 2008. "Scaling Regression Inputs by Dividing by Two Standard Deviations." Statistics in Medicine 27: 2865--2873.

Examples

data.frame(x = cumsum(rnorm(10)),
           y = cumsum(rnorm(10)),
           group = sample(c("A", "B"))) -> Example

# Note: this works in a group_by() context, which you should do in a case like this.
diff_at(Example, c("x", "y"))
#>             x         y group        d_x         d_y
#> 1   1.0373702 -1.947543     A         NA          NA
#> 2   1.9922077 -2.754220     B  0.9548375 -0.80667721
#> 3  -0.5138991 -3.716241     A -2.5061067 -0.96202125
#> 4  -1.4374897 -3.705119     B -0.9235907  0.01112254
#> 5  -1.1472210 -4.150026     A  0.2902687 -0.44490690
#> 6  -0.6408681 -4.230861     B  0.5063529 -0.08083509
#> 7  -1.5424786 -4.946060     A -0.9016106 -0.71519918
#> 8  -2.0308380 -3.980065     B -0.4883593  0.96599493
#> 9  -2.9490707 -5.787184     A -0.9182328 -1.80711920
#> 10 -2.4172062 -6.003793     B  0.5318645 -0.21660873

r1sd_at(Example, c("x", "y"))
#>             x         y group        s_x         s_y
#> 1   1.0373702 -1.947543     A  1.3160218  1.74581979
#> 2   1.9922077 -2.754220     B  1.9435875  1.09819100
#> 3  -0.5138991 -3.716241     A  0.2964522  0.32584656
#> 4  -1.4374897 -3.705119     B -0.3105765  0.33477613
#> 5  -1.1472210 -4.150026     A -0.1197978 -0.02241075
#> 6  -0.6408681 -4.230861     B  0.2130020 -0.08730800
#> 7  -1.5424786 -4.946060     A -0.3795804 -0.66149502
#> 8  -2.0308380 -3.980065     B -0.7005539  0.11403963
#> 9  -2.9490707 -5.787184     A -1.3040611 -1.33677912
#> 10 -2.4172062 -6.003793     B -0.9544939 -1.51068022

r2sd_at(Example, c("x", "y"))
#>             x         y group        z_x         z_y
#> 1   1.0373702 -1.947543     A  0.6580109  0.87290989
#> 2   1.9922077 -2.754220     B  0.9717938  0.54909550
#> 3  -0.5138991 -3.716241     A  0.1482261  0.16292328
#> 4  -1.4374897 -3.705119     B -0.1552883  0.16738806
#> 5  -1.1472210 -4.150026     A -0.0598989 -0.01120538
#> 6  -0.6408681 -4.230861     B  0.1065010 -0.04365400
#> 7  -1.5424786 -4.946060     A -0.1897902 -0.33074751
#> 8  -2.0308380 -3.980065     B -0.3502769  0.05701981
#> 9  -2.9490707 -5.787184     A -0.6520306 -0.66838956
#> 10 -2.4172062 -6.003793     B -0.4772469 -0.75534011