Create multivariate data by permutation

corvectors() is a function to obtain a multivariate dataset by specifying the relation between those specified variables.

corvectors(
  data,
  corm,
  tol = 0.005,
  conv = 10000,
  cores = 2,
  splitsize = 1000,
  verbose = FALSE,
  seed
)

Arguments

data
a data matrix containing the data

corm
A value containing the desired correlation or a vector or data matrix containing the desired correlations

tol
A single value or a vector of tolerances with length ncol(data) - 1. The default is 0.005

conv
The maximum iterations allowed. Defaults to 1000.

cores
The number of cores to be used for parallel computing

splitsize
The size to use for splitting the data

verbose
Logical statement. Default is FALSE

seed
An optional seed to set

Value

corvectors() returns a matrix given the specified multivariate relation.

Details

This is liberally copy-pasted from van Kooten and Vink's wonderful-but-no-longer-supported correlate package. They call it correlate() in their package, but I opt for corvectors() here.

Author

Pascal van Kooten and Gerko Vink

Examples

# \donttest{
# bivariate example, start with zero correlation
corvectors(replicate(2, rnorm(100)), .5)

#>               [,1]        [,2]
#>   [1,] -1.23713996  0.78295517
#>   [2,] -1.37798670 -0.52564039
#>   [3,]  0.27503126  1.09260573
#>   [4,] -0.66795544 -0.05013828
#>   [5,]  0.55738252 -0.76755218
#>   [6,] -1.01238384 -1.74999650
#>   [7,] -0.79544564 -2.27522072
#>   [8,] -0.27931928 -0.30465688
#>   [9,]  0.11678369 -0.65190423
#>  [10,]  1.26771349  0.62760451
#>  [11,]  0.85023624  2.74885198
#>  [12,]  1.08873598 -0.07750108
#>  [13,]  0.63667947 -0.03211670
#>  [14,]  0.69005635  0.88528647
#>  [15,] -0.50127757 -1.15436535
#>  [16,]  0.39607245  0.23607535
#>  [17,]  0.21533313 -0.24417342
#>  [18,]  0.05717782 -0.30940725
#>  [19,] -0.21208884 -0.59044355
#>  [20,]  1.14885697  0.58007782
#>  [21,]  0.58584400  0.12069883
#>  [22,]  0.02838126 -0.23104140
#>  [23,]  1.24298865  0.40015727
#>  [24,] -1.16394046 -2.68897992
#>  [25,] -0.80488027 -0.19232924
#>  [26,] -0.94238578 -1.77461569
#>  [27,]  0.95223303  1.78104065
#>  [28,]  0.18068492 -0.85451732
#>  [29,]  0.40534701 -0.19944165
#>  [30,]  0.97439470  0.52698723
#>  [31,]  0.57382730 -0.22757969
#>  [32,] -0.06665350  1.67444832
#>  [33,] -1.64928906 -1.12558527
#>  [34,] -0.88329743 -0.49884358
#>  [35,] -0.02985419 -1.99372143
#>  [36,] -1.32556083  0.50315772
#>  [37,] -1.43988413 -1.34523643
#>  [38,] -0.57787399 -0.46607721
#>  [39,]  0.82465469  0.73471324
#>  [40,]  0.96041726  0.70001274
#>  [41,] -0.09757365  0.58415928
#>  [42,] -1.80303954  0.63420154
#>  [43,]  0.87752485  1.47291735
#>  [44,]  0.57964236  2.82811446
#>  [45,] -0.22191503  1.40618777
#>  [46,]  0.15710255  0.68882635
#>  [47,]  2.21357740  0.58784717
#>  [48,]  1.53677480  1.37146581
#>  [49,] -0.38173896 -0.44799849
#>  [50,]  0.36841078  0.09252502
#>  [51,]  0.83242230  1.34044840
#>  [52,] -0.05342496 -0.39862179
#>  [53,]  1.40223138  0.90039651
#>  [54,]  2.30398889 -1.21288296
#>  [55,]  0.09183975 -0.54074261
#>  [56,]  0.40128442  0.20946459
#>  [57,]  0.87577954  0.22893182
#>  [58,] -0.46277729  0.67368064
#>  [59,] -1.34454748 -0.64475668
#>  [60,]  0.71258675 -0.36960776
#>  [61,] -0.18013489  0.74526138
#>  [62,]  0.94812875 -0.01523648
#>  [63,] -1.47331528 -1.65526740
#>  [64,] -1.01059002 -0.57851390
#>  [65,]  1.47606428  1.66760165
#>  [66,]  1.15357665  0.05235419
#>  [67,]  0.42370435 -0.46623541
#>  [68,]  0.12733224  2.44064189
#>  [69,]  0.20401676  1.27335615
#>  [70,] -0.87586766  0.36985875
#>  [71,] -1.81887642 -2.42406367
#>  [72,]  0.86323943  0.55382939
#>  [73,] -0.95543146 -0.59279926
#>  [74,]  0.48464660  0.18838184
#>  [75,]  0.07063484 -0.33489310
#>  [76,] -1.62339383  0.66873200
#>  [77,] -1.19123927 -0.95706074
#>  [78,]  0.73428443 -0.48212424
#>  [79,] -1.62721113 -1.16226101
#>  [80,]  0.69523628 -0.05673319
#>  [81,] -0.59425060  0.15758900
#>  [82,]  0.75867769  1.55156748
#>  [83,] -0.86264027  0.15639488
#>  [84,]  0.14214254 -1.27624821
#>  [85,] -0.83372758 -0.45845865
#>  [86,]  0.59378781  1.33375112
#>  [87,]  0.74541620  0.43600571
#>  [88,]  0.15931532  0.20483938
#>  [89,] -0.27178909 -0.58454153
#>  [90,]  0.64325952  0.31243750
#>  [91,]  1.84601226  1.17282506
#>  [92,]  0.05349542  0.01691176
#>  [93,]  1.12966650  2.84297711
#>  [94,] -1.83473442 -2.09110033
#>  [95,] -0.72324656 -1.25214027
#>  [96,] -1.29654378 -0.38345966
#>  [97,] -2.03005826 -0.50103961
#>  [98,] -1.11557458  0.23723810
#>  [99,]  1.81787943  0.17915543
#> [100,] -0.60372181 -1.83357608

# multivariate example

corvectors(replicate(4, rnorm(100)), c(.5, .6, .7))

#>                [,1]        [,2]          [,3]        [,4]
#>   [1,]  2.644207881 -1.04351781  0.5362566452  0.30612905
#>   [2,] -0.492397255 -0.84282368 -0.2104967419 -0.32525890
#>   [3,]  1.190919496  1.05753055 -0.3274316434  1.31831936
#>   [4,]  0.011435379  0.33123534 -1.3265596845  0.56191433
#>   [5,] -0.994484454 -0.16523844 -0.9672961182 -0.99550872
#>   [6,] -0.265780702 -0.40828894 -0.2126370201 -0.57663880
#>   [7,]  0.410887352 -0.56398978 -0.0910310274  0.58400667
#>   [8,]  0.569655600  0.39441911 -2.0437915398  0.18170594
#>   [9,] -1.116461415 -0.29747219 -1.5686406094 -2.36862038
#>  [10,]  0.876834224 -0.20731681  0.4335510968  0.54816134
#>  [11,] -1.375407315  0.09019305  0.4061677289 -0.61154601
#>  [12,] -1.160520585 -0.78157737 -0.8455728641 -0.68910646
#>  [13,] -0.690722001 -2.11029480 -1.4599688995  0.03753717
#>  [14,]  0.240004183 -0.49125533  0.3848072394 -0.94185589
#>  [15,]  0.808817471 -0.41235823 -0.0577985765 -0.50328651
#>  [16,]  1.051770247  0.47004143  0.7755412508  1.12770630
#>  [17,] -0.220592316 -1.08123631  1.3289726614 -0.08117769
#>  [18,] -0.589970576 -0.84408829 -0.5456140286  0.03836601
#>  [19,]  0.913474182  0.87823360  0.2565472463  1.30167123
#>  [20,] -0.322511877 -1.20972138 -0.1060391712 -0.61669720
#>  [21,]  0.290481577 -1.39678434 -0.7149206208  0.37413219
#>  [22,]  2.296067925  1.18029960  0.4822020837  0.37074333
#>  [23,]  1.359367943  0.18098977  0.8286152849  2.24264440
#>  [24,] -1.216507239 -1.01659737 -0.2356272164 -0.36733317
#>  [25,]  0.398420568  0.35930351 -0.6418591737  0.79507291
#>  [26,] -0.723500629 -1.19789594 -3.0894352360 -0.62901425
#>  [27,]  0.952518055  0.44392701  0.6116691758  1.84900220
#>  [28,]  0.368889548 -0.39847044 -0.1240873481  0.79158308
#>  [29,] -1.359494302 -1.00461589 -2.2255636442  0.44691751
#>  [30,]  0.854473180 -0.14298005  0.8152939461  0.49176899
#>  [31,]  0.856400688  0.77918774  1.1742713573  1.08974505
#>  [32,] -1.309596288  0.18513027 -0.9538672719 -0.43691114
#>  [33,]  0.357329417  0.54088533  0.1487794597 -0.64983275
#>  [34,] -0.245251176 -0.98647138  0.5773073256 -0.30199837
#>  [35,] -0.791961485 -0.78455851 -1.1882322835 -0.07374903
#>  [36,]  0.324243941  0.07568805  0.0003354162  0.13370272
#>  [37,]  1.069327650  1.51192885  0.1680478492  0.60726051
#>  [38,] -0.278886128  0.20485154  0.3008584329 -0.20268002
#>  [39,] -1.505210671 -0.41468476 -1.2418506418 -1.63546723
#>  [40,]  0.069963489 -0.36188198  0.6698955476  0.02153617
#>  [41,]  0.370505323  2.63702198 -0.4691859180  0.81224776
#>  [42,]  0.188761060 -0.36606413  0.2415867947  0.07450573
#>  [43,] -1.413843556 -1.10455689 -2.3493743763 -0.82768915
#>  [44,] -0.170133717 -0.38778273 -0.8970311528  1.00198235
#>  [45,] -1.714500828 -1.35655088 -0.5943657383 -2.66390252
#>  [46,]  0.304431589 -0.25060297  0.6521673355  1.14721768
#>  [47,]  0.329035317 -1.00253494  0.5835806337  1.51657069
#>  [48,]  0.792254194  0.69944094  0.5095251013  0.78325213
#>  [49,] -1.022800171 -0.75049498 -1.4286526082 -1.15882355
#>  [50,]  1.697780094  1.63518955  0.5535638207  1.92581982
#>  [51,]  1.959219540  2.46158496  0.7627511726  0.59454180
#>  [52,] -0.064879340  0.97970469  0.1333107197  0.49903189
#>  [53,] -0.649725571 -1.22299757 -0.4592829994  0.73542438
#>  [54,] -0.329182942 -1.59015583 -0.3566609226 -0.05690401
#>  [55,]  2.400336315 -0.10930596  1.0479527641  0.74854692
#>  [56,]  0.446771850 -0.17256468  1.6235174760  0.85816471
#>  [57,]  0.919900677  1.05219768  0.5975467169  1.17214317
#>  [58,]  0.502463635  1.02382334  1.9643038332  0.31225666
#>  [59,]  0.995493455 -0.36742194  0.2325732736  1.37696889
#>  [60,] -0.003242194  0.88210078  0.3799635345 -0.45299585
#>  [61,]  0.960209539  0.53259254  0.8707731056  1.27810397
#>  [62,] -2.006770573 -1.57479453 -1.1463802331 -2.59893783
#>  [63,] -1.022034554 -0.61871197  0.9403737083 -1.43256541
#>  [64,] -0.144444311 -0.37456531  0.1774606137  0.36017218
#>  [65,]  0.573236409  0.59881664  1.0509399192  0.16177237
#>  [66,]  0.417664788  1.32272035  0.9629150348  1.65117237
#>  [67,] -0.529243516  1.65971118 -1.5131367373 -0.20298343
#>  [68,]  0.700780971 -0.67671648  0.2181233494  1.66686308
#>  [69,] -1.139776506 -1.37138600 -1.5581079182  0.31079782
#>  [70,]  0.083485819  0.57937085  0.0375011369 -0.20031125
#>  [71,] -0.493967588  0.02721972  0.6118348662 -1.05970756
#>  [72,]  0.416942603 -0.27329385 -0.8079519619  0.81244842
#>  [73,] -0.518899067 -0.74744442 -0.9096406402  0.64735028
#>  [74,] -0.638373988  0.32668682 -0.6728544658 -0.15387397
#>  [75,] -0.614170631 -1.62894187  0.2859219307 -0.68608953
#>  [76,] -1.248588219  1.32307524 -1.6254394597 -0.69973530
#>  [77,] -0.372598426 -0.60652492 -0.9408416011  0.51496278
#>  [78,] -0.173367132  0.52275573 -0.0261211558 -0.99359893
#>  [79,]  1.317589020  0.86776719  0.2907359833  1.01878641
#>  [80,]  1.177362306  0.80615393  1.6973770978  2.55104456
#>  [81,]  0.448064957  0.33233584  0.1878873330  0.59697047
#>  [82,]  0.375397197 -0.21921244  2.0191359215  0.32591825
#>  [83,] -0.890155673  0.18152016 -0.4669223683 -0.55662979
#>  [84,] -0.681955338 -0.72877434 -1.0097065526 -1.37314045
#>  [85,]  0.434144729 -1.01701460  0.9225958161  2.01164066
#>  [86,]  0.525751122  0.45335177  0.7002846833  1.12767045
#>  [87,] -1.005007825 -1.25310923 -0.5862873480 -1.30526405
#>  [88,] -0.030234395 -0.14907846 -0.8308306171 -0.20380532
#>  [89,] -0.443932644 -0.53761370  0.3177095208  0.15500160
#>  [90,] -0.417372961  0.92470161  0.5334002829 -0.30826867
#>  [91,] -0.580833650 -0.95745550 -0.7662563744 -0.26542525
#>  [92,]  1.507944168 -0.21338344  1.5018116681  1.28645452
#>  [93,]  0.053264927 -0.07944460  0.3750084287  0.56476553
#>  [94,] -0.573338347 -0.61289399 -0.8655774730 -1.08754992
#>  [95,]  1.207273774  1.94609717  1.9179735966  1.72208083
#>  [96,] -0.841458972  0.18176087 -0.9340483852  0.84174985
#>  [97,]  1.460464713  0.54979738  0.2168132359  1.34212122
#>  [98,] -0.482080212 -1.40657270 -1.1163058225 -0.53651222
#>  [99,]  0.610472638  0.58039429  2.4998958986  1.49694733
#> [100,]  1.759922223  1.14507853  0.4494456385  0.60673796

# }