Scale a vector by two standard deviations

r2sd allows you to rescale a numeric vector such that the ensuing output has a mean of 0 and a standard deviation of .5.

r2sd(x, na = TRUE)

Arguments

x
a numeric vector

na
what to do with NAs in the vector. Defaults to TRUE (i.e. passes over the missing observations)

Value

The function returns a numeric vector rescaled with a mean of 0 and a standard deviation of .5.

Details

By default, `na.rm` is set to TRUE. If you have missing data, the function will just pass over them.

Gelman (2008) argues that rescaling by two standard deviations puts regression inputs on roughly the same scale no matter their original scale. This allows for some honest, if preliminary, assessment of relative effect sizes from the regression output. This does that, but without requiring the rescale function from arm. I'm trying to reduce the packages on which my workflow relies.

Importantly, I tend to rescale only the ordinal and interval inputs and leave the binary inputs as 0/1. So, my r2sd function doesn't have any of the fancier if-else statements that Gelman's rescale function has.

References

Gelman, Andrew. 2008. "Scaling Regression Inputs by Dividing by Two Standard Deviations." Statistics in Medicine 27: 2865--2873.

Examples

x <- rnorm(100)
r2sd(x)

#>   [1]  0.58938435  0.18197583  0.58039130  0.91074730  0.01668217 -0.18478770
#>   [7] -0.01962033 -0.42179684  0.36908293  0.06542483 -0.88627356  0.28892837
#>  [13] -0.98614536  0.50553860 -0.51143024 -0.23288148  0.31638425 -0.16144326
#>  [19] -0.07254027  1.07923937  1.02523837  0.21173516  0.01595284  0.21346861
#>  [25] -0.36107322 -0.05857578 -1.27614690 -0.35787795 -0.22397223  0.15244818
#>  [31] -0.23264380  0.95027701 -0.03677793 -0.54346030  0.43066237 -0.39212414
#>  [37] -0.58999446  0.58785329 -0.08272961 -0.05357957  0.17797965  0.60034952
#>  [43] -0.25598906  0.04699418  0.15419367  0.58142698  0.74035987  0.56813722
#>  [49] -0.09961741  0.03155173  0.36507027  0.24989286  0.13771068 -0.08150362
#>  [55]  0.40591857 -0.22073909  1.06251295 -0.14558226 -0.53391304 -0.61590666
#>  [61]  0.03739685  0.00918478 -0.24785673 -0.34481666  0.39803186 -1.28439892
#>  [67]  0.42355915 -0.14919976  0.53538693  0.43795067  0.25980651  0.23653250
#>  [73] -0.31081525 -0.53525793 -0.02476267  0.27148311 -1.27299036 -0.55505474
#>  [79] -0.67254477 -0.12399040 -1.14078240 -0.22691123 -0.64409311  0.43743081
#>  [85]  0.15814365 -0.09573691 -0.19889374 -0.21807596  0.06221853  0.08736401
#>  [91]  0.59476649 -0.18209050  0.70475613  0.33387888  0.25281224 -0.22382176
#>  [97]  0.29469543  0.21196985 -0.64322852 -0.62643329