An Incomplete List of My Publications, All of Which You Should Cite

These are data on my publications, barring a few things like book reviews and some forthcoming pieces. I use these data to illustrate the print_refs() function. You should cite my publications more.

Usage

stevepubs

Format

A data frame with the following 14 variables.

CATEGORY

the entry type

BIBTEXKEY

the unique entry key

AUTHOR

a list of authors for this entry

BOOKTITLE

the book title, if appropriate

JOURNAL

the journal title, if appropriate

NUMBER

the journal volume number, if appropriate

PAGES

the range of page numbers, if appropriate

PUBLISHER

the book publisher, if appropriate

TITLE

the title of the publication

VOLUME

the journal volume number, if appropriate

YEAR

the year of publication, as a character. Publications with no year are assumed to be forthcoming

DOI

a DOI, if I entered one

Details

Cite my publications more, you goons. Extremely Smokey Bear voice Only YOU can jack my h-index to infinity.