Scoped Helper Verbs

Scoped helper verbs included in this R Documentation file allow for targeted commands on specified columns. They also rename the ensuing output to conform to my preferred style. The commands here are multiple and explained in the details section below.

center_at(data, x, prefix = "c", na = TRUE)

diff_at(data, x, l = 1, prefix = "d")

group_mean_center_at(data, x, mean_prefix = "mean", prefix = "b", na = TRUE)

log_at(data, x, prefix = "ln", plus_1 = FALSE)

mean_at(data, x, prefix = "mean", na = TRUE)

r1sd_at(data, x, prefix = "s", na = TRUE)

r2sd_at(data, x, prefix = "z", na = TRUE)

Arguments

data

a data frame

x

a vector, likely in your data frame

prefix

Allows the user to rename the prefix of the new variables. Each function has defaults (see details section).

na

what to do with NAs in the vector. Defaults to TRUE (i.e. passes over the missing observations). Not applicable to diff_at.

l

The order of lags for calculating differences in diff_at. Applicable only to diff_at.

mean_prefix

Applicable only to group_mean_center_at. Specifies the prefix of the (assumed) total population mean variables. Default is "mean", though the user can change this as they see fit.

plus_1

Applicable only to log_at. If TRUE, adds 1 to the variables prior to log transformation. If FALSE, performs logarithmic transformation on variables no matter whether 0 occurs (i.e. 0s will come back as -Inf). Defaults to FALSE.

Value

The function returns a set of new vectors in a data frame after performing relevant functions. The new vectors have distinct prefixes corresponding with the action performed on them.

Details

center_at is a wrapper for mutate_at and rename_at from dplyr. It takes supplied vectors and effectively centers them from the mean. It then renames these new variables to have a prefix of c_. The default prefix ("c") can be changed by way of an argument in the function.

diff_at is a wrapper for mutate_at and rename_at from dplyr. It takes supplied vectors and creates differences from the previous value recorded above it. It then renames these new variables to have a prefix of d_ (in the case of a first difference), or something like d2_ in the case of second differences, or d3_ in the case of third differences (and so on). The exact prefix depends on the l argument, which communicates the order of lags you want. It defaults to 1. The default prefix ("d") can be changed by way of an argument in the function.

group_mean_center_at is a wrapper for mutate and across in dplyr. It takes supplied vectors and centers an (assumed) group mean of the variables from an (assumed) total population mean of the variables provided to it. It then returns the new variables with a prefix, whose default is b_. This prefix communicates, if you will, a kind of "between" variable in the panel model context, in juxtaposition to "within" variables in the panel model context.

log_at is a wrapper for mutate_at and rename_at from dplyr. It takes supplied vectors and creates a variable that takes a natural logarithmic transformation of them. It then renames these new variables to have a prefix of ln_. This default prefix ("ln") can be changed by way of an argument in the function.

mean_at is a wrapper for mutate_at and rename_at from dplyr. It takes supplied vectors and creates a variable communicating the mean of the variable. It then renames these new variables to have a prefix of mean_. This default prefix ("mean") can be changed by way of an argument in the function.

r1sd_at is a wrapper for mutate_at and rename_at from dplyr. It both rescales the supplied vectors to new vectors and renames the vectors to each have a prefix of s_. Note the rescaling here is just by one standard deviation and not two. The default prefix ("s") can be changed by way of an argument in the function.

r2sd_at is a wrapper for mutate_at and rename_at from dplyr. It both rescales the supplied vectors to new vectors and renames the vectors to each have a prefix of z_. Note the rescaling here is by two standard deviations and not one. The default prefix ("z") can be changed by way of an argument in the function.

These functions will fail in the absence of a character vector of a length of one. In which case, you should think about using some other dplyr verb on its own.

Examples

Example <- data.frame(category = c(rep("A", 5),
                                   rep("B", 5),
                                   rep("C", 5)),
                      x = runif(15), y = runif(15))

center_at(Example, c("x", "y"))
#>    category          x          y         c_x        c_y
#> 1         A 0.74672576 0.41752823  0.21340519 -0.1493581
#> 2         A 0.14856527 0.83598232 -0.38475530  0.2690960
#> 3         A 0.89587575 0.18192990  0.36255518 -0.3849564
#> 4         A 0.39631973 0.39844948 -0.13700084 -0.1684368
#> 5         A 0.87384983 0.34417597  0.34052926 -0.2227103
#> 6         B 0.43550339 0.73398220 -0.09781719  0.1670959
#> 7         B 0.60518333 0.96748654  0.07186275  0.4006002
#> 8         B 0.02544965 0.95474160 -0.50787092  0.3878553
#> 9         B 0.80846803 0.13886299  0.27514745 -0.4280233
#> 10        B 0.64031172 0.99922778  0.10699115  0.4323415
#> 11        C 0.25293613 0.83591651 -0.28038445  0.2690302
#> 12        C 0.73604217 0.08408537  0.20272160 -0.4828009
#> 13        C 0.10544946 0.85101375 -0.42787112  0.2841275
#> 14        C 0.37755757 0.08334624 -0.15576300 -0.4835401
#> 15        C 0.95157082 0.67656550  0.41825024  0.1096792

diff_at(Example, c("x", "y"))
#>    category          x          y        d_x         d_y
#> 1         A 0.74672576 0.41752823         NA          NA
#> 2         A 0.14856527 0.83598232 -0.5981605  0.41845409
#> 3         A 0.89587575 0.18192990  0.7473105 -0.65405243
#> 4         A 0.39631973 0.39844948 -0.4995560  0.21651958
#> 5         A 0.87384983 0.34417597  0.4775301 -0.05427351
#> 6         B 0.43550339 0.73398220 -0.4383464  0.38980623
#> 7         B 0.60518333 0.96748654  0.1696799  0.23350434
#> 8         B 0.02544965 0.95474160 -0.5797337 -0.01274494
#> 9         B 0.80846803 0.13886299  0.7830184 -0.81587861
#> 10        B 0.64031172 0.99922778 -0.1681563  0.86036479
#> 11        C 0.25293613 0.83591651 -0.3873756 -0.16331128
#> 12        C 0.73604217 0.08408537  0.4831060 -0.75183113
#> 13        C 0.10544946 0.85101375 -0.6305927  0.76692837
#> 14        C 0.37755757 0.08334624  0.2721081 -0.76766751
#> 15        C 0.95157082 0.67656550  0.5740132  0.59321926

diff_at(Example, c("x", "y"), l=3)
#>    category          x          y        d3_x        d3_y
#> 1         A 0.74672576 0.41752823          NA          NA
#> 2         A 0.14856527 0.83598232          NA          NA
#> 3         A 0.89587575 0.18192990          NA          NA
#> 4         A 0.39631973 0.39844948 -0.35040603 -0.01907875
#> 5         A 0.87384983 0.34417597  0.72528456 -0.49180635
#> 6         B 0.43550339 0.73398220 -0.46037236  0.55205230
#> 7         B 0.60518333 0.96748654  0.20886360  0.56903706
#> 8         B 0.02544965 0.95474160 -0.84840018  0.61056564
#> 9         B 0.80846803 0.13886299  0.37296464 -0.59511920
#> 10        B 0.64031172 0.99922778  0.03512839  0.03174124
#> 11        C 0.25293613 0.83591651  0.22748647 -0.11882510
#> 12        C 0.73604217 0.08408537 -0.07242585 -0.05477762
#> 13        C 0.10544946 0.85101375 -0.53486227 -0.14821404
#> 14        C 0.37755757 0.08334624  0.12462144 -0.75257027
#> 15        C 0.95157082 0.67656550  0.21552864  0.59248012

log_at(Example, c("x", "y"))
#>    category          x          y        ln_x          ln_y
#> 1         A 0.74672576 0.41752823 -0.29205728 -0.8734031252
#> 2         A 0.14856527 0.83598232 -1.90673088 -0.1791478112
#> 3         A 0.89587575 0.18192990 -0.10995355 -1.7041338471
#> 4         A 0.39631973 0.39844948 -0.92553399 -0.9201745734
#> 5         A 0.87384983 0.34417597 -0.13484673 -1.0666022173
#> 6         B 0.43550339 0.73398220 -0.83125271 -0.3092705029
#> 7         B 0.60518333 0.96748654 -0.50222385 -0.0330537650
#> 8         B 0.02544965 0.95474160 -3.67105319 -0.0463145467
#> 9         B 0.80846803 0.13886299 -0.21261415 -1.9742674830
#> 10        B 0.64031172 0.99922778 -0.44580016 -0.0007725154
#> 11        C 0.25293613 0.83591651 -1.37461828 -0.1792265439
#> 12        C 0.73604217 0.08408537 -0.30646786 -2.4759226319
#> 13        C 0.10544946 0.85101375 -2.24952353 -0.1613269960
#> 14        C 0.37755757 0.08334624 -0.97403222 -2.4847518210
#> 15        C 0.95157082 0.67656550 -0.04964117 -0.3907260214

log_at(Example, c("x", "y"), plus_1 = TRUE)
#>    category          x          y      ln_x       ln_y
#> 1         A 0.74672576 0.41752823 0.5577430 0.34891467
#> 2         A 0.14856527 0.83598232 0.1385136 0.60757966
#> 3         A 0.89587575 0.18192990 0.6396809 0.16714861
#> 4         A 0.39631973 0.39844948 0.3338400 0.33536411
#> 5         A 0.87384983 0.34417597 0.6279950 0.29578116
#> 6         B 0.43550339 0.73398220 0.3615156 0.55042061
#> 7         B 0.60518333 0.96748654 0.4732380 0.67675686
#> 8         B 0.02544965 0.95474160 0.0251312 0.67025801
#> 9         B 0.80846803 0.13886299 0.5924801 0.13003039
#> 10        B 0.64031172 0.99922778 0.4948863 0.69276100
#> 11        C 0.25293613 0.83591651 0.2254897 0.60754382
#> 12        C 0.73604217 0.08408537 0.5516079 0.08073666
#> 13        C 0.10544946 0.85101375 0.1002520 0.61573346
#> 14        C 0.37755757 0.08334624 0.3203121 0.08005462
#> 15        C 0.95157082 0.67656550 0.6686346 0.51674735

mean_at(Example, c("x", "y"))
#>    category          x          y    mean_x    mean_y
#> 1         A 0.74672576 0.41752823 0.5333206 0.5668863
#> 2         A 0.14856527 0.83598232 0.5333206 0.5668863
#> 3         A 0.89587575 0.18192990 0.5333206 0.5668863
#> 4         A 0.39631973 0.39844948 0.5333206 0.5668863
#> 5         A 0.87384983 0.34417597 0.5333206 0.5668863
#> 6         B 0.43550339 0.73398220 0.5333206 0.5668863
#> 7         B 0.60518333 0.96748654 0.5333206 0.5668863
#> 8         B 0.02544965 0.95474160 0.5333206 0.5668863
#> 9         B 0.80846803 0.13886299 0.5333206 0.5668863
#> 10        B 0.64031172 0.99922778 0.5333206 0.5668863
#> 11        C 0.25293613 0.83591651 0.5333206 0.5668863
#> 12        C 0.73604217 0.08408537 0.5333206 0.5668863
#> 13        C 0.10544946 0.85101375 0.5333206 0.5668863
#> 14        C 0.37755757 0.08334624 0.5333206 0.5668863
#> 15        C 0.95157082 0.67656550 0.5333206 0.5668863

group_mean_center_at(mean_at(Example, c("x", "y")), c("x", "y"))
#> Warning: The absence of a grouping factor means you're going to be subtracting a mean from itself. It's up to you if you want to do that. You probably didn't want that, but I did it anyway. I'm just bringing that to your attention.
#> Warning: The ensuing output for these variables you created are all 0, indicating you subtracted a thing from a different version of itself in this function. Are you sure these mean variables aren't themselves actually group means? Check your code more carefully.
#>    category          x          y    mean_x    mean_y b_x b_y
#> 1         A 0.74672576 0.41752823 0.5333206 0.5668863   0   0
#> 2         A 0.14856527 0.83598232 0.5333206 0.5668863   0   0
#> 3         A 0.89587575 0.18192990 0.5333206 0.5668863   0   0
#> 4         A 0.39631973 0.39844948 0.5333206 0.5668863   0   0
#> 5         A 0.87384983 0.34417597 0.5333206 0.5668863   0   0
#> 6         B 0.43550339 0.73398220 0.5333206 0.5668863   0   0
#> 7         B 0.60518333 0.96748654 0.5333206 0.5668863   0   0
#> 8         B 0.02544965 0.95474160 0.5333206 0.5668863   0   0
#> 9         B 0.80846803 0.13886299 0.5333206 0.5668863   0   0
#> 10        B 0.64031172 0.99922778 0.5333206 0.5668863   0   0
#> 11        C 0.25293613 0.83591651 0.5333206 0.5668863   0   0
#> 12        C 0.73604217 0.08408537 0.5333206 0.5668863   0   0
#> 13        C 0.10544946 0.85101375 0.5333206 0.5668863   0   0
#> 14        C 0.37755757 0.08334624 0.5333206 0.5668863   0   0
#> 15        C 0.95157082 0.67656550 0.5333206 0.5668863   0   0
# ^ Alternatively, a better way:
# Example %>%
#   mean_at(c("x", "y")) %>%
#   group_by(category) %>%
#   group_mean_center_at(c("x", "y"))

r1sd_at(Example, c("x", "y"))
#>    category          x          y        s_x        s_y
#> 1         A 0.74672576 0.41752823  0.6945870 -0.4335035
#> 2         A 0.14856527 0.83598232 -1.2522940  0.7810363
#> 3         A 0.89587575 0.18192990  1.1800375 -1.1173145
#> 4         A 0.39631973 0.39844948 -0.4459077 -0.4888785
#> 5         A 0.87384983 0.34417597  1.1083480 -0.6464043
#> 6         B 0.43550339 0.73398220 -0.3183735  0.4849866
#> 7         B 0.60518333 0.96748654  0.2338975  1.1627199
#> 8         B 0.02544965 0.95474160 -1.6530083  1.1257285
#> 9         B 0.80846803 0.13886299  0.8955445 -1.2423138
#> 10        B 0.64031172 0.99922778  0.3482327  1.2548471
#> 11        C 0.25293613 0.83591651 -0.9125898  0.7808452
#> 12        C 0.73604217 0.08408537  0.6598143 -1.4013028
#> 13        C 0.10544946 0.85101375 -1.3926265  0.8246641
#> 14        C 0.37755757 0.08334624 -0.5069744 -1.4034481
#> 15        C 0.95157082 0.67656550  1.3613127  0.3183378

r2sd_at(Example, c("x", "y"))
#>    category          x          y        z_x        z_y
#> 1         A 0.74672576 0.41752823  0.3472935 -0.2167517
#> 2         A 0.14856527 0.83598232 -0.6261470  0.3905181
#> 3         A 0.89587575 0.18192990  0.5900187 -0.5586573
#> 4         A 0.39631973 0.39844948 -0.2229538 -0.2444392
#> 5         A 0.87384983 0.34417597  0.5541740 -0.3232022
#> 6         B 0.43550339 0.73398220 -0.1591867  0.2424933
#> 7         B 0.60518333 0.96748654  0.1169487  0.5813600
#> 8         B 0.02544965 0.95474160 -0.8265042  0.5628642
#> 9         B 0.80846803 0.13886299  0.4477723 -0.6211569
#> 10        B 0.64031172 0.99922778  0.1741163  0.6274236
#> 11        C 0.25293613 0.83591651 -0.4562949  0.3904226
#> 12        C 0.73604217 0.08408537  0.3299071 -0.7006514
#> 13        C 0.10544946 0.85101375 -0.6963133  0.4123321
#> 14        C 0.37755757 0.08334624 -0.2534872 -0.7017241
#> 15        C 0.95157082 0.67656550  0.6806563  0.1591689