Scoped Helper Verbs

Scoped helper verbs included in this R Documentation file allow for targeted commands on specified columns. They also rename the ensuing output to conform to my preferred style. The commands here are multiple and explained in the details section below.

center_at(data, x, na = TRUE)

diff_at(data, x, l = 1)

r1sd_at(data, x, na = TRUE)

r2sd_at(data, x, na = TRUE)

within_at(data, x, na = TRUE)

Arguments

data

a data frame

x

a vector, likely in your data frame

na

what to do with NAs in the vector. Defaults to TRUE (i.e. passes over the missing observations). Not applicable to diff_at.

l

The order of lags for calculating differences in diff_at. Applicable only to diff_at.

Value

The function returns a set of new vectors in a data frame after performing relevant functions. The new vectors have distinct prefixes corresponding with the action performed on them.

Details

center_at is a wrapper for mutate_at and rename_at from dplyr. It takes supplied vectors and effectively centers them from the mean. It then renames these new variables to have a prefix of c_. It is best used without a group_by preceding it. The underlying code is identical to within_at in this same package, though it results in new variables with a different prefix.

r1sd_at is a wrapper for mutate_at and rename_at from dplyr. It both rescales the supplied vectors to new vectors and renames the vectors to each have a prefix of s_. Note the rescaling here is just by one standard deviation and not two.

r2sd_at is a wrapper for mutate_at and rename_at from dplyr. It both rescales the supplied vectors to new vectors and renames the vectors to each have a prefix of z_. Note the rescaling here is by two standard deviations and not one.

within_at is a wrapper for mutate_at and rename_at from dplyr. It takes supplied vectors and effectively centers them from the mean. It then renames these new variables to have a prefix of w_. It is best used with a group_by preceding it on some kind of "subject", with an eye toward ultimately including them as covariates in a panel model to isolate so-called "within" effects.

Examples

center_at(USArrests, c("Murder", "Assault"))
#>                Murder Assault UrbanPop Rape c_Murder c_Assault
#> Alabama          13.2     236       58 21.2    5.412     65.24
#> Alaska           10.0     263       48 44.5    2.212     92.24
#> Arizona           8.1     294       80 31.0    0.312    123.24
#> Arkansas          8.8     190       50 19.5    1.012     19.24
#> California        9.0     276       91 40.6    1.212    105.24
#> Colorado          7.9     204       78 38.7    0.112     33.24
#> Connecticut       3.3     110       77 11.1   -4.488    -60.76
#> Delaware          5.9     238       72 15.8   -1.888     67.24
#> Florida          15.4     335       80 31.9    7.612    164.24
#> Georgia          17.4     211       60 25.8    9.612     40.24
#> Hawaii            5.3      46       83 20.2   -2.488   -124.76
#> Idaho             2.6     120       54 14.2   -5.188    -50.76
#> Illinois         10.4     249       83 24.0    2.612     78.24
#> Indiana           7.2     113       65 21.0   -0.588    -57.76
#> Iowa              2.2      56       57 11.3   -5.588   -114.76
#> Kansas            6.0     115       66 18.0   -1.788    -55.76
#> Kentucky          9.7     109       52 16.3    1.912    -61.76
#> Louisiana        15.4     249       66 22.2    7.612     78.24
#> Maine             2.1      83       51  7.8   -5.688    -87.76
#> Maryland         11.3     300       67 27.8    3.512    129.24
#> Massachusetts     4.4     149       85 16.3   -3.388    -21.76
#> Michigan         12.1     255       74 35.1    4.312     84.24
#> Minnesota         2.7      72       66 14.9   -5.088    -98.76
#> Mississippi      16.1     259       44 17.1    8.312     88.24
#> Missouri          9.0     178       70 28.2    1.212      7.24
#> Montana           6.0     109       53 16.4   -1.788    -61.76
#> Nebraska          4.3     102       62 16.5   -3.488    -68.76
#> Nevada           12.2     252       81 46.0    4.412     81.24
#> New Hampshire     2.1      57       56  9.5   -5.688   -113.76
#> New Jersey        7.4     159       89 18.8   -0.388    -11.76
#> New Mexico       11.4     285       70 32.1    3.612    114.24
#> New York         11.1     254       86 26.1    3.312     83.24
#> North Carolina   13.0     337       45 16.1    5.212    166.24
#> North Dakota      0.8      45       44  7.3   -6.988   -125.76
#> Ohio              7.3     120       75 21.4   -0.488    -50.76
#> Oklahoma          6.6     151       68 20.0   -1.188    -19.76
#> Oregon            4.9     159       67 29.3   -2.888    -11.76
#> Pennsylvania      6.3     106       72 14.9   -1.488    -64.76
#> Rhode Island      3.4     174       87  8.3   -4.388      3.24
#> South Carolina   14.4     279       48 22.5    6.612    108.24
#> South Dakota      3.8      86       45 12.8   -3.988    -84.76
#> Tennessee        13.2     188       59 26.9    5.412     17.24
#> Texas            12.7     201       80 25.5    4.912     30.24
#> Utah              3.2     120       80 22.9   -4.588    -50.76
#> Vermont           2.2      48       32 11.2   -5.588   -122.76
#> Virginia          8.5     156       63 20.7    0.712    -14.76
#> Washington        4.0     145       73 26.2   -3.788    -25.76
#> West Virginia     5.7      81       39  9.3   -2.088    -89.76
#> Wisconsin         2.6      53       66 10.8   -5.188   -117.76
#> Wyoming           6.8     161       60 15.6   -0.988     -9.76

diff_at(USArrests, c("Murder", "Assault"))
#>                Murder Assault UrbanPop Rape d_Murder d_Assault
#> Alabama          13.2     236       58 21.2       NA        NA
#> Alaska           10.0     263       48 44.5     -3.2        27
#> Arizona           8.1     294       80 31.0     -1.9        31
#> Arkansas          8.8     190       50 19.5      0.7      -104
#> California        9.0     276       91 40.6      0.2        86
#> Colorado          7.9     204       78 38.7     -1.1       -72
#> Connecticut       3.3     110       77 11.1     -4.6       -94
#> Delaware          5.9     238       72 15.8      2.6       128
#> Florida          15.4     335       80 31.9      9.5        97
#> Georgia          17.4     211       60 25.8      2.0      -124
#> Hawaii            5.3      46       83 20.2    -12.1      -165
#> Idaho             2.6     120       54 14.2     -2.7        74
#> Illinois         10.4     249       83 24.0      7.8       129
#> Indiana           7.2     113       65 21.0     -3.2      -136
#> Iowa              2.2      56       57 11.3     -5.0       -57
#> Kansas            6.0     115       66 18.0      3.8        59
#> Kentucky          9.7     109       52 16.3      3.7        -6
#> Louisiana        15.4     249       66 22.2      5.7       140
#> Maine             2.1      83       51  7.8    -13.3      -166
#> Maryland         11.3     300       67 27.8      9.2       217
#> Massachusetts     4.4     149       85 16.3     -6.9      -151
#> Michigan         12.1     255       74 35.1      7.7       106
#> Minnesota         2.7      72       66 14.9     -9.4      -183
#> Mississippi      16.1     259       44 17.1     13.4       187
#> Missouri          9.0     178       70 28.2     -7.1       -81
#> Montana           6.0     109       53 16.4     -3.0       -69
#> Nebraska          4.3     102       62 16.5     -1.7        -7
#> Nevada           12.2     252       81 46.0      7.9       150
#> New Hampshire     2.1      57       56  9.5    -10.1      -195
#> New Jersey        7.4     159       89 18.8      5.3       102
#> New Mexico       11.4     285       70 32.1      4.0       126
#> New York         11.1     254       86 26.1     -0.3       -31
#> North Carolina   13.0     337       45 16.1      1.9        83
#> North Dakota      0.8      45       44  7.3    -12.2      -292
#> Ohio              7.3     120       75 21.4      6.5        75
#> Oklahoma          6.6     151       68 20.0     -0.7        31
#> Oregon            4.9     159       67 29.3     -1.7         8
#> Pennsylvania      6.3     106       72 14.9      1.4       -53
#> Rhode Island      3.4     174       87  8.3     -2.9        68
#> South Carolina   14.4     279       48 22.5     11.0       105
#> South Dakota      3.8      86       45 12.8    -10.6      -193
#> Tennessee        13.2     188       59 26.9      9.4       102
#> Texas            12.7     201       80 25.5     -0.5        13
#> Utah              3.2     120       80 22.9     -9.5       -81
#> Vermont           2.2      48       32 11.2     -1.0       -72
#> Virginia          8.5     156       63 20.7      6.3       108
#> Washington        4.0     145       73 26.2     -4.5       -11
#> West Virginia     5.7      81       39  9.3      1.7       -64
#> Wisconsin         2.6      53       66 10.8     -3.1       -28
#> Wyoming           6.8     161       60 15.6      4.2       108

r1sd_at(USArrests, c("Murder", "Assault"))
#>                Murder Assault UrbanPop Rape    s_Murder   s_Assault
#> Alabama          13.2     236       58 21.2  1.24256408  0.78283935
#> Alaska           10.0     263       48 44.5  0.50786248  1.10682252
#> Arizona           8.1     294       80 31.0  0.07163341  1.47880321
#> Arkansas          8.8     190       50 19.5  0.23234938  0.23086801
#> California        9.0     276       91 40.6  0.27826823  1.26281442
#> Colorado          7.9     204       78 38.7  0.02571456  0.39885929
#> Connecticut       3.3     110       77 11.1 -1.03041900 -0.72908214
#> Delaware          5.9     238       72 15.8 -0.43347395  0.80683810
#> Florida          15.4     335       80 31.9  1.74767144  1.97077766
#> Georgia          17.4     211       60 25.8  2.20685994  0.48285493
#> Hawaii            5.3      46       83 20.2 -0.57123050 -1.49704226
#> Idaho             2.6     120       54 14.2 -1.19113497 -0.60908837
#> Illinois         10.4     249       83 24.0  0.59970018  0.93883125
#> Indiana           7.2     113       65 21.0 -0.13500142 -0.69308401
#> Iowa              2.2      56       57 11.3 -1.28297267 -1.37704849
#> Kansas            6.0     115       66 18.0 -0.41051452 -0.66908525
#> Kentucky          9.7     109       52 16.3  0.43898421 -0.74108152
#> Louisiana        15.4     249       66 22.2  1.74767144  0.93883125
#> Maine             2.1      83       51  7.8 -1.30593210 -1.05306531
#> Maryland         11.3     300       67 27.8  0.80633501  1.55079947
#> Massachusetts     4.4     149       85 16.3 -0.77786532 -0.26110644
#> Michigan         12.1     255       74 35.1  0.99001041  1.01082751
#> Minnesota         2.7      72       66 14.9 -1.16817555 -1.18505846
#> Mississippi      16.1     259       44 17.1  1.90838741  1.05882502
#> Missouri          9.0     178       70 28.2  0.27826823  0.08687549
#> Montana           6.0     109       53 16.4 -0.41051452 -0.74108152
#> Nebraska          4.3     102       62 16.5 -0.80082475 -0.82507715
#> Nevada           12.2     252       81 46.0  1.01296983  0.97482938
#> New Hampshire     2.1      57       56  9.5 -1.30593210 -1.36504911
#> New Jersey        7.4     159       89 18.8 -0.08908257 -0.14111267
#> New Mexico       11.4     285       70 32.1  0.82929443  1.37080881
#> New York         11.1     254       86 26.1  0.76041616  0.99882813
#> North Carolina   13.0     337       45 16.1  1.19664523  1.99477641
#> North Dakota      0.8      45       44  7.3 -1.60440462 -1.50904164
#> Ohio              7.3     120       75 21.4 -0.11204199 -0.60908837
#> Oklahoma          6.6     151       68 20.0 -0.27275797 -0.23710769
#> Oregon            4.9     159       67 29.3 -0.66306820 -0.14111267
#> Pennsylvania      6.3     106       72 14.9 -0.34163624 -0.77707965
#> Rhode Island      3.4     174       87  8.3 -1.00745957  0.03887798
#> South Carolina   14.4     279       48 22.5  1.51807718  1.29881255
#> South Dakota      3.8      86       45 12.8 -0.91562187 -1.01706718
#> Tennessee        13.2     188       59 26.9  1.24256408  0.20686926
#> Texas            12.7     201       80 25.5  1.12776696  0.36286116
#> Utah              3.2     120       80 22.9 -1.05337842 -0.60908837
#> Vermont           2.2      48       32 11.2 -1.28297267 -1.47304350
#> Virginia          8.5     156       63 20.7  0.16347111 -0.17711080
#> Washington        4.0     145       73 26.2 -0.86970302 -0.30910395
#> West Virginia     5.7      81       39  9.3 -0.47939280 -1.07706407
#> Wisconsin         2.6      53       66 10.8 -1.19113497 -1.41304662
#> Wyoming           6.8     161       60 15.6 -0.22683912 -0.11711392

r2sd_at(USArrests, c("Murder", "Assault"))
#>                Murder Assault UrbanPop Rape    z_Murder   z_Assault
#> Alabama          13.2     236       58 21.2  0.62128204  0.39141967
#> Alaska           10.0     263       48 44.5  0.25393124  0.55341126
#> Arizona           8.1     294       80 31.0  0.03581670  0.73940160
#> Arkansas          8.8     190       50 19.5  0.11617469  0.11543401
#> California        9.0     276       91 40.6  0.13913412  0.63140721
#> Colorado          7.9     204       78 38.7  0.01285728  0.19942964
#> Connecticut       3.3     110       77 11.1 -0.51520950 -0.36454107
#> Delaware          5.9     238       72 15.8 -0.21673697  0.40341905
#> Florida          15.4     335       80 31.9  0.87383572  0.98538883
#> Georgia          17.4     211       60 25.8  1.10342997  0.24142746
#> Hawaii            5.3      46       83 20.2 -0.28561525 -0.74852113
#> Idaho             2.6     120       54 14.2 -0.59556749 -0.30454418
#> Illinois         10.4     249       83 24.0  0.29985009  0.46941562
#> Indiana           7.2     113       65 21.0 -0.06750071 -0.34654200
#> Iowa              2.2      56       57 11.3 -0.64148634 -0.68852424
#> Kansas            6.0     115       66 18.0 -0.20525726 -0.33454263
#> Kentucky          9.7     109       52 16.3  0.21949210 -0.37054076
#> Louisiana        15.4     249       66 22.2  0.87383572  0.46941562
#> Maine             2.1      83       51  7.8 -0.65296605 -0.52653266
#> Maryland         11.3     300       67 27.8  0.40316750  0.77539973
#> Massachusetts     4.4     149       85 16.3 -0.38893266 -0.13055322
#> Michigan         12.1     255       74 35.1  0.49500520  0.50541375
#> Minnesota         2.7      72       66 14.9 -0.58408777 -0.59252923
#> Mississippi      16.1     259       44 17.1  0.95419371  0.52941251
#> Missouri          9.0     178       70 28.2  0.13913412  0.04343774
#> Montana           6.0     109       53 16.4 -0.20525726 -0.37054076
#> Nebraska          4.3     102       62 16.5 -0.40041237 -0.41253858
#> Nevada           12.2     252       81 46.0  0.50648492  0.48741469
#> New Hampshire     2.1      57       56  9.5 -0.65296605 -0.68252456
#> New Jersey        7.4     159       89 18.8 -0.04454128 -0.07055634
#> New Mexico       11.4     285       70 32.1  0.41464722  0.68540441
#> New York         11.1     254       86 26.1  0.38020808  0.49941407
#> North Carolina   13.0     337       45 16.1  0.59832262  0.99738821
#> North Dakota      0.8      45       44  7.3 -0.80220231 -0.75452082
#> Ohio              7.3     120       75 21.4 -0.05602100 -0.30454418
#> Oklahoma          6.6     151       68 20.0 -0.13637898 -0.11855384
#> Oregon            4.9     159       67 29.3 -0.33153410 -0.07055634
#> Pennsylvania      6.3     106       72 14.9 -0.17081812 -0.38853982
#> Rhode Island      3.4     174       87  8.3 -0.50372979  0.01943899
#> South Carolina   14.4     279       48 22.5  0.75903859  0.64940628
#> South Dakota      3.8      86       45 12.8 -0.45781094 -0.50853359
#> Tennessee        13.2     188       59 26.9  0.62128204  0.10343463
#> Texas            12.7     201       80 25.5  0.56388348  0.18143058
#> Utah              3.2     120       80 22.9 -0.52668921 -0.30454418
#> Vermont           2.2      48       32 11.2 -0.64148634 -0.73652175
#> Virginia          8.5     156       63 20.7  0.08173555 -0.08855540
#> Washington        4.0     145       73 26.2 -0.43485151 -0.15455197
#> West Virginia     5.7      81       39  9.3 -0.23969640 -0.53853203
#> Wisconsin         2.6      53       66 10.8 -0.59556749 -0.70652331
#> Wyoming           6.8     161       60 15.6 -0.11341956 -0.05855696

# Note: best used with a group_by() preceding it.
within_at(USArrests, c("Murder", "Assault"))
#>                Murder Assault UrbanPop Rape w_Murder w_Assault
#> Alabama          13.2     236       58 21.2    5.412     65.24
#> Alaska           10.0     263       48 44.5    2.212     92.24
#> Arizona           8.1     294       80 31.0    0.312    123.24
#> Arkansas          8.8     190       50 19.5    1.012     19.24
#> California        9.0     276       91 40.6    1.212    105.24
#> Colorado          7.9     204       78 38.7    0.112     33.24
#> Connecticut       3.3     110       77 11.1   -4.488    -60.76
#> Delaware          5.9     238       72 15.8   -1.888     67.24
#> Florida          15.4     335       80 31.9    7.612    164.24
#> Georgia          17.4     211       60 25.8    9.612     40.24
#> Hawaii            5.3      46       83 20.2   -2.488   -124.76
#> Idaho             2.6     120       54 14.2   -5.188    -50.76
#> Illinois         10.4     249       83 24.0    2.612     78.24
#> Indiana           7.2     113       65 21.0   -0.588    -57.76
#> Iowa              2.2      56       57 11.3   -5.588   -114.76
#> Kansas            6.0     115       66 18.0   -1.788    -55.76
#> Kentucky          9.7     109       52 16.3    1.912    -61.76
#> Louisiana        15.4     249       66 22.2    7.612     78.24
#> Maine             2.1      83       51  7.8   -5.688    -87.76
#> Maryland         11.3     300       67 27.8    3.512    129.24
#> Massachusetts     4.4     149       85 16.3   -3.388    -21.76
#> Michigan         12.1     255       74 35.1    4.312     84.24
#> Minnesota         2.7      72       66 14.9   -5.088    -98.76
#> Mississippi      16.1     259       44 17.1    8.312     88.24
#> Missouri          9.0     178       70 28.2    1.212      7.24
#> Montana           6.0     109       53 16.4   -1.788    -61.76
#> Nebraska          4.3     102       62 16.5   -3.488    -68.76
#> Nevada           12.2     252       81 46.0    4.412     81.24
#> New Hampshire     2.1      57       56  9.5   -5.688   -113.76
#> New Jersey        7.4     159       89 18.8   -0.388    -11.76
#> New Mexico       11.4     285       70 32.1    3.612    114.24
#> New York         11.1     254       86 26.1    3.312     83.24
#> North Carolina   13.0     337       45 16.1    5.212    166.24
#> North Dakota      0.8      45       44  7.3   -6.988   -125.76
#> Ohio              7.3     120       75 21.4   -0.488    -50.76
#> Oklahoma          6.6     151       68 20.0   -1.188    -19.76
#> Oregon            4.9     159       67 29.3   -2.888    -11.76
#> Pennsylvania      6.3     106       72 14.9   -1.488    -64.76
#> Rhode Island      3.4     174       87  8.3   -4.388      3.24
#> South Carolina   14.4     279       48 22.5    6.612    108.24
#> South Dakota      3.8      86       45 12.8   -3.988    -84.76
#> Tennessee        13.2     188       59 26.9    5.412     17.24
#> Texas            12.7     201       80 25.5    4.912     30.24
#> Utah              3.2     120       80 22.9   -4.588    -50.76
#> Vermont           2.2      48       32 11.2   -5.588   -122.76
#> Virginia          8.5     156       63 20.7    0.712    -14.76
#> Washington        4.0     145       73 26.2   -3.788    -25.76
#> West Virginia     5.7      81       39  9.3   -2.088    -89.76
#> Wisconsin         2.6      53       66 10.8   -5.188   -117.76
#> Wyoming           6.8     161       60 15.6   -0.988     -9.76