Scoped Helper Verbs

Scoped helper verbs included in this R Documentation file allow for targeted commands on specified columns. They also rename the ensuing output to conform to my preferred style. The commands here are multiple and explained in the details section below.

center_at(data, x, prefix = "c", na = TRUE)

diff_at(data, x, l = 1, prefix = "d")

group_mean_center_at(data, x, mean_prefix = "mean", prefix = "b", na = TRUE)

mean_at(data, x, prefix = "mean", na = TRUE)

r1sd_at(data, x, prefix = "s", na = TRUE)

r2sd_at(data, x, prefix = "z", na = TRUE)

Arguments

data

a data frame

x

a vector, likely in your data frame

prefix

Allows the user to rename the prefix of the new variables. Each function has defaults (see details section).

na

what to do with NAs in the vector. Defaults to TRUE (i.e. passes over the missing observations). Not applicable to diff_at.

l

The order of lags for calculating differences in diff_at. Applicable only to diff_at.

mean_prefix

Applicable only to group_mean_center_at. Specifies the prefix of the (assumed) total population mean variables. Default is "mean", though the user can change this as they see fit.

Value

The function returns a set of new vectors in a data frame after performing relevant functions. The new vectors have distinct prefixes corresponding with the action performed on them.

Details

center_at is a wrapper for mutate_at and rename_at from dplyr. It takes supplied vectors and effectively centers them from the mean. It then renames these new variables to have a prefix of c_. The default prefix ("c") can be changed by way of an argument in the function.

diff_at is a wrapper for mutate_at and rename_at from dplyr. It takes supplied vectors and creates differences from the previous value recorded above it. It then renames these new variables to have a prefix of d_ (in the case of a first difference), or something like d2_ in the case of second differences, or d3_ in the case of third differences (and so on). The exact prefix depends on the l argument, which communicates the order of lags you want. It defaults to 1. The default prefix ("d") can be changed by way of an argument in the function.

group_mean_center_at is a wrapper for mutate and across in dplyr. It takes supplied vectors and centers an (assumed) group mean of the variables from an (assumed) total population mean of the variables provided to it. It then returns the new variables with a prefix, whose default is b_. This prefix communicates, if you will, a kind of "between" variable in the panel model context, in juxtaposition to "within" variables in the panel model context.

mean_at is a wrapper for mutate_at and rename_at from dplyr. It takes supplied vectors and creates a variable communicating the mean of the variable. It then renames these new variables to have a prefix of mean_. This default prefix ("mean") can be changed by way of an argument in the function.

r1sd_at is a wrapper for mutate_at and rename_at from dplyr. It both rescales the supplied vectors to new vectors and renames the vectors to each have a prefix of s_. Note the rescaling here is just by one standard deviation and not two. The default prefix ("s") can be changed by way of an argument in the function.

r2sd_at is a wrapper for mutate_at and rename_at from dplyr. It both rescales the supplied vectors to new vectors and renames the vectors to each have a prefix of z_. Note the rescaling here is by two standard deviations and not one. The default prefix ("z") can be changed by way of an argument in the function.

These functions will fail in the absence of a character vector of a length of one. In which case, you should think about using some other dplyr verb on its own.

Examples

Example <- data.frame(category = c(rep("A", 5),
                                   rep("B", 5),
                                   rep("C", 5)),
                      x = runif(15), y = rnorm(15))

center_at(Example, c("x", "y"))
#>    category           x          y         c_x         c_y
#> 1         A 0.527996932  0.7511701  0.06548559  0.96366745
#> 2         A 0.108754827 -0.1140987 -0.35375651  0.09839867
#> 3         A 0.367758927 -0.9281261 -0.09475241 -0.71562874
#> 4         A 0.156813549 -0.9328369 -0.30569779 -0.72033962
#> 5         A 0.724172281 -0.4621337  0.26166094 -0.24963636
#> 6         B 0.141979856 -0.7497709 -0.32053148 -0.53727362
#> 7         B 0.356398022 -0.6013288 -0.10611332 -0.38883143
#> 8         B 0.557475245  0.1242021  0.09496390  0.33669938
#> 9         B 0.006035023 -0.2071348 -0.45647632  0.00536249
#> 10        B 0.563462667 -1.4801478  0.10095133 -1.26765048
#> 11        C 0.807182184  0.1080437  0.34467084  0.32054099
#> 12        C 0.388162453  0.3945454 -0.07434889  0.60704273
#> 13        C 0.516811815 -0.3752977  0.05430048 -0.16280036
#> 14        C 0.833782667  1.1758299  0.37127133  1.38832718
#> 15        C 0.880883655  0.1096244  0.41837231  0.32212171

diff_at(Example, c("x", "y"))
#>    category           x          y         d_x          d_y
#> 1         A 0.527996932  0.7511701          NA           NA
#> 2         A 0.108754827 -0.1140987 -0.41924210 -0.865268772
#> 3         A 0.367758927 -0.9281261  0.25900410 -0.814027410
#> 4         A 0.156813549 -0.9328369 -0.21094538 -0.004710879
#> 5         A 0.724172281 -0.4621337  0.56735873  0.470703259
#> 6         B 0.141979856 -0.7497709 -0.58219242 -0.287637264
#> 7         B 0.356398022 -0.6013288  0.21441817  0.148442187
#> 8         B 0.557475245  0.1242021  0.20107722  0.725530817
#> 9         B 0.006035023 -0.2071348 -0.55144022 -0.331336893
#> 10        B 0.563462667 -1.4801478  0.55742764 -1.273012966
#> 11        C 0.807182184  0.1080437  0.24371952  1.588191466
#> 12        C 0.388162453  0.3945454 -0.41901973  0.286501734
#> 13        C 0.516811815 -0.3752977  0.12864936 -0.769843090
#> 14        C 0.833782667  1.1758299  0.31697085  1.551127546
#> 15        C 0.880883655  0.1096244  0.04710099 -1.066205474

diff_at(Example, c("x", "y"), l=3)
#>    category           x          y        d3_x        d3_y
#> 1         A 0.527996932  0.7511701          NA          NA
#> 2         A 0.108754827 -0.1140987          NA          NA
#> 3         A 0.367758927 -0.9281261          NA          NA
#> 4         A 0.156813549 -0.9328369 -0.37118338 -1.68400706
#> 5         A 0.724172281 -0.4621337  0.61541745 -0.34803503
#> 6         B 0.141979856 -0.7497709 -0.22577907  0.17835512
#> 7         B 0.356398022 -0.6013288  0.19958447  0.33150818
#> 8         B 0.557475245  0.1242021 -0.16669704  0.58633574
#> 9         B 0.006035023 -0.2071348 -0.13594483  0.54263611
#> 10        B 0.563462667 -1.4801478  0.20706465 -0.87881904
#> 11        C 0.807182184  0.1080437  0.24970694 -0.01615839
#> 12        C 0.388162453  0.3945454  0.38212743  0.60168023
#> 13        C 0.516811815 -0.3752977 -0.04665085  1.10485011
#> 14        C 0.833782667  1.1758299  0.02660048  1.06778619
#> 15        C 0.880883655  0.1096244  0.49272120 -0.28492102

mean_at(Example, c("x", "y"))
#>    category           x          y    mean_x     mean_y
#> 1         A 0.527996932  0.7511701 0.4625113 -0.2124973
#> 2         A 0.108754827 -0.1140987 0.4625113 -0.2124973
#> 3         A 0.367758927 -0.9281261 0.4625113 -0.2124973
#> 4         A 0.156813549 -0.9328369 0.4625113 -0.2124973
#> 5         A 0.724172281 -0.4621337 0.4625113 -0.2124973
#> 6         B 0.141979856 -0.7497709 0.4625113 -0.2124973
#> 7         B 0.356398022 -0.6013288 0.4625113 -0.2124973
#> 8         B 0.557475245  0.1242021 0.4625113 -0.2124973
#> 9         B 0.006035023 -0.2071348 0.4625113 -0.2124973
#> 10        B 0.563462667 -1.4801478 0.4625113 -0.2124973
#> 11        C 0.807182184  0.1080437 0.4625113 -0.2124973
#> 12        C 0.388162453  0.3945454 0.4625113 -0.2124973
#> 13        C 0.516811815 -0.3752977 0.4625113 -0.2124973
#> 14        C 0.833782667  1.1758299 0.4625113 -0.2124973
#> 15        C 0.880883655  0.1096244 0.4625113 -0.2124973

group_mean_center_at(mean_at(Example, c("x", "y")), c("x", "y"))
#> Warning: The absence of a grouping factor means you're going to be subtracting a mean from itself. It's up to you if you want to do that. You probably didn't want that, but I did it anyway. I'm just bringing that to your attention.
#>    category           x          y    mean_x     mean_y b_x b_y
#> 1         A 0.527996932  0.7511701 0.4625113 -0.2124973   0   0
#> 2         A 0.108754827 -0.1140987 0.4625113 -0.2124973   0   0
#> 3         A 0.367758927 -0.9281261 0.4625113 -0.2124973   0   0
#> 4         A 0.156813549 -0.9328369 0.4625113 -0.2124973   0   0
#> 5         A 0.724172281 -0.4621337 0.4625113 -0.2124973   0   0
#> 6         B 0.141979856 -0.7497709 0.4625113 -0.2124973   0   0
#> 7         B 0.356398022 -0.6013288 0.4625113 -0.2124973   0   0
#> 8         B 0.557475245  0.1242021 0.4625113 -0.2124973   0   0
#> 9         B 0.006035023 -0.2071348 0.4625113 -0.2124973   0   0
#> 10        B 0.563462667 -1.4801478 0.4625113 -0.2124973   0   0
#> 11        C 0.807182184  0.1080437 0.4625113 -0.2124973   0   0
#> 12        C 0.388162453  0.3945454 0.4625113 -0.2124973   0   0
#> 13        C 0.516811815 -0.3752977 0.4625113 -0.2124973   0   0
#> 14        C 0.833782667  1.1758299 0.4625113 -0.2124973   0   0
#> 15        C 0.880883655  0.1096244 0.4625113 -0.2124973   0   0
# ^ Alternatively, a better way:
# Example %>%
#   mean_at(c("x", "y")) %>%
#   group_by(category) %>%
#   group_mean_center_at(c("x", "y"))

r1sd_at(Example, c("x", "y"))
#>    category           x          y        s_x          s_y
#> 1         A 0.527996932  0.7511701  0.2355505  1.390849179
#> 2         A 0.108754827 -0.1140987 -1.2724560  0.142017577
#> 3         A 0.367758927 -0.9281261 -0.3408228 -1.032858009
#> 4         A 0.156813549 -0.9328369 -1.0995896 -1.039657162
#> 5         A 0.724172281 -0.4621337  0.9411898 -0.360297032
#> 6         B 0.141979856 -0.7497709 -1.1529461 -0.775440300
#> 7         B 0.356398022 -0.6013288 -0.3816877 -0.561195547
#> 8         B 0.557475245  0.1242021  0.3415835  0.485954011
#> 9         B 0.006035023 -0.2071348 -1.6419373  0.007739615
#> 10        B 0.563462667 -1.4801478  0.3631202 -1.829584085
#> 11        C 0.807182184  0.1080437  1.2397750  0.462632806
#> 12        C 0.388162453  0.3945454 -0.2674316  0.876137177
#> 13        C 0.516811815 -0.3752977  0.1953179 -0.234967731
#> 14        C 0.833782667  1.1758299  1.3354565  2.003755266
#> 15        C 0.880883655  0.1096244  1.5048779  0.464914233

r2sd_at(Example, c("x", "y"))
#>    category           x          y         z_x          z_y
#> 1         A 0.527996932  0.7511701  0.11777526  0.695424590
#> 2         A 0.108754827 -0.1140987 -0.63622799  0.071008788
#> 3         A 0.367758927 -0.9281261 -0.17041138 -0.516429004
#> 4         A 0.156813549 -0.9328369 -0.54979480 -0.519828581
#> 5         A 0.724172281 -0.4621337  0.47059491 -0.180148516
#> 6         B 0.141979856 -0.7497709 -0.57647306 -0.387720150
#> 7         B 0.356398022 -0.6013288 -0.19084387 -0.280597773
#> 8         B 0.557475245  0.1242021  0.17079175  0.242977006
#> 9         B 0.006035023 -0.2071348 -0.82096865  0.003869808
#> 10        B 0.563462667 -1.4801478  0.18156008 -0.914792042
#> 11        C 0.807182184  0.1080437  0.61988749  0.231316403
#> 12        C 0.388162453  0.3945454 -0.13371582  0.438068589
#> 13        C 0.516811815 -0.3752977  0.09765893 -0.117483865
#> 14        C 0.833782667  1.1758299  0.66772823  1.001877633
#> 15        C 0.880883655  0.1096244  0.75243894  0.232457117