Create multivariate data by permutation

corvectors() is a function to obtain a multivariate dataset by specifying the relation between those specified variables.

corvectors(
  data,
  corm,
  tol = 0.005,
  conv = 10000,
  cores = 2,
  splitsize = 1000,
  verbose = FALSE,
  seed
)

Arguments

data
a data matrix containing the data

corm
A value containing the desired correlation or a vector or data matrix containing the desired correlations

tol
A single value or a vector of tolerances with length ncol(data) - 1. The default is 0.005

conv
The maximum iterations allowed. Defaults to 1000.

cores
The number of cores to be used for parallel computing

splitsize
The size to use for splitting the data

verbose
Logical statement. Default is FALSE

seed
An optional seed to set

Value

corvectors() returns a matrix given the specified multivariate relation.

Details

This is liberally copy-pasted from van Kooten and Vink's wonderful-but-no-longer-supported correlate package. They call it correlate() in their package, but I opt for corvectors() here.

Author

Pascal van Kooten and Gerko Vink

Examples

# \donttest{
# bivariate example, start with zero correlation
corvectors(replicate(2, rnorm(100)), .5)

#>               [,1]        [,2]
#>   [1,]  0.36841078  0.88528647
#>   [2,] -1.81887642 -2.09110033
#>   [3,]  0.63667947  1.09260573
#>   [4,]  1.26771349  0.62760451
#>   [5,] -1.62721113 -1.16226101
#>   [6,]  0.59378781  1.33375112
#>   [7,]  0.07063484 -0.33489310
#>   [8,] -0.27178909  1.17282506
#>   [9,]  0.85023624  0.73471324
#>  [10,] -0.95543146 -0.54074261
#>  [11,] -0.50127757  0.12069883
#>  [12,]  1.24298865  0.40015727
#>  [13,] -1.19123927 -0.36960776
#>  [14,]  0.05717782 -0.30940725
#>  [15,]  0.71258675 -0.50103961
#>  [16,]  1.53677480 -0.44799849
#>  [17,]  0.11678369 -0.59044355
#>  [18,]  0.64325952  0.31243750
#>  [19,]  0.73428443 -0.48212424
#>  [20,]  0.15931532  0.20483938
#>  [21,]  1.08873598  1.47291735
#>  [22,]  0.75867769  0.15758900
#>  [23,] -0.86264027 -0.19232924
#>  [24,] -1.32556083 -0.76755218
#>  [25,]  1.40223138  0.17915543
#>  [26,]  0.02838126 -0.64475668
#>  [27,]  1.14885697  2.82811446
#>  [28,]  0.21533313 -1.12558527
#>  [29,]  0.69005635 -1.15436535
#>  [30,] -1.83473442  0.55382939
#>  [31,] -1.01059002  0.20946459
#>  [32,] -0.05342496 -0.39862179
#>  [33,]  0.74541620  0.43600571
#>  [34,] -0.27931928 -0.30465688
#>  [35,] -0.06665350 -1.77461569
#>  [36,]  0.27503126 -0.05673319
#>  [37,] -1.11557458 -0.58454153
#>  [38,]  0.69523628 -0.03211670
#>  [39,]  1.81787943 -0.38345966
#>  [40,]  1.12966650  2.44064189
#>  [41,] -0.22191503 -0.52564039
#>  [42,]  0.40534701 -0.22757969
#>  [43,]  0.57964236  1.37146581
#>  [44,] -1.16394046 -2.68897992
#>  [45,] -1.43988413 -1.34523643
#>  [46,] -0.02985419 -1.99372143
#>  [47,]  0.82465469 -0.23104140
#>  [48,]  0.12733224  0.23723810
#>  [49,]  0.87752485 -0.07750108
#>  [50,]  2.21357740  2.74885198
#>  [51,] -0.46277729  0.67368064
#>  [52,] -0.83372758 -0.45845865
#>  [53,] -1.64928906 -0.65190423
#>  [54,]  0.86323943 -1.21288296
#>  [55,]  0.57382730  0.68882635
#>  [56,] -1.80303954  0.63420154
#>  [57,] -2.03005826 -0.46623541
#>  [58,] -1.34454748  0.01691176
#>  [59,]  0.39607245  0.23607535
#>  [60,]  0.09183975  0.90039651
#>  [61,]  0.96041726  0.70001274
#>  [62,] -0.09757365  0.58415928
#>  [63,] -1.47331528 -1.65526740
#>  [64,]  0.83242230 -0.01523648
#>  [65,]  0.15710255 -1.27624821
#>  [66,] -0.38173896 -0.46607721
#>  [67,] -0.66795544 -1.74999650
#>  [68,] -0.79544564 -0.85451732
#>  [69,] -1.62339383 -0.95706074
#>  [70,]  0.14214254  0.18838184
#>  [71,] -1.29654378 -0.59279926
#>  [72,]  0.20401676  1.27335615
#>  [73,] -0.80488027  0.15639488
#>  [74,] -1.23713996  0.78295517
#>  [75,] -0.18013489  0.74526138
#>  [76,]  0.94812875  1.55156748
#>  [77,] -0.60372181  0.66873200
#>  [78,]  2.30398889  0.36985875
#>  [79,]  1.47606428  1.66760165
#>  [80,] -0.87586766 -2.42406367
#>  [81,]  0.58584400  0.52698723
#>  [82,] -0.72324656 -1.25214027
#>  [83,]  0.40128442 -0.57851390
#>  [84,]  0.48464660  0.58784717
#>  [85,]  0.95223303  1.78104065
#>  [86,]  0.55738252  0.50315772
#>  [87,] -0.59425060 -1.83357608
#>  [88,]  1.15357665  0.05235419
#>  [89,] -0.21208884 -0.24417342
#>  [90,]  0.97439470  1.40618777
#>  [91,]  1.84601226  2.84297711
#>  [92,]  0.42370435  0.22893182
#>  [93,] -1.01238384 -0.05013828
#>  [94,]  0.18068492  0.09252502
#>  [95,]  0.87577954  1.34044840
#>  [96,] -1.37798670 -0.19944165
#>  [97,]  0.05349542  0.58007782
#>  [98,] -0.94238578  1.67444832
#>  [99,] -0.57787399 -2.27522072
#> [100,] -0.88329743 -0.49884358

# multivariate example

corvectors(replicate(4, rnorm(100)), c(.5, .6, .7))

#>               [,1]         [,2]        [,3]         [,4]
#>   [1,]  1.26909874  1.966005813 -0.13526123  0.854729795
#>   [2,]  0.97689574  0.656268934 -0.79795890  0.838498070
#>   [3,] -2.18148220 -0.871343534 -2.63476217 -1.230426427
#>   [4,] -1.54045159 -0.821814282  0.40940301 -1.056952930
#>   [5,]  0.15582633  0.287029289  0.86146747  0.240003410
#>   [6,]  0.55960799  1.130206248  0.53609713 -1.108072673
#>   [7,]  0.57329175  0.754664124  1.15907409  0.877864728
#>   [8,] -0.56474842  0.154456625 -1.08689716 -0.108257385
#>   [9,] -1.27159377 -1.943743970 -1.44527138 -0.881994772
#>  [10,]  0.10058908 -0.953251686  0.43456184 -0.148987739
#>  [11,]  0.21173985  0.754171788 -1.29440313 -0.055816850
#>  [12,]  1.64452950 -0.186635638  0.03483027  1.007735914
#>  [13,] -0.02138997  0.298739488 -0.68443284  0.464356645
#>  [14,]  2.10009938  2.053432056  0.83720866  1.444298911
#>  [15,]  0.98970979  1.106520200  0.18874710  1.028794837
#>  [16,] -0.61270142 -1.417567541 -0.59656571 -1.189192439
#>  [17,] -0.26526440 -0.531498259 -0.49424668  0.004547368
#>  [18,] -0.69175285  0.030415815 -1.96584642 -0.766504690
#>  [19,] -0.58820519 -0.067817580 -0.54814577  0.286447469
#>  [20,]  0.11584263 -0.174536631  1.41297085  0.065326041
#>  [21,] -0.05391672  1.781155182  0.07773320  0.010487419
#>  [22,] -1.14127536 -0.745794384 -1.07101514 -0.864550017
#>  [23,] -1.03943719 -0.994638485 -1.00308699  0.065956326
#>  [24,]  0.73928097  0.295488516  0.30754431  0.538204556
#>  [25,]  1.54984808  1.585682889  1.31094764  0.235275193
#>  [26,]  0.14064843 -0.935318498 -0.37032900 -0.039513516
#>  [27,]  1.76184976  0.848906847  1.60107048  1.305091034
#>  [28,]  0.24099362  0.315088754 -0.20938254 -0.250056521
#>  [29,]  0.85685890  0.498325898  1.00575203  0.853891811
#>  [30,]  0.80198951  2.722736198  0.85063603  1.208294558
#>  [31,] -0.89492084  0.276742211 -1.00849282 -1.244151318
#>  [32,]  1.63313589  1.422454361  1.79917534  0.341189200
#>  [33,] -1.88518794 -0.007078339 -0.99656870 -1.130457353
#>  [34,]  0.52509316 -1.901881647  1.89836407 -0.049541490
#>  [35,] -1.10691482 -1.552024799 -0.78173922 -0.579249553
#>  [36,] -1.02393310  0.052164947 -0.95836237 -0.496935015
#>  [37,]  0.04411055  0.699658467  0.28839845  0.077184053
#>  [38,]  0.78793034  0.725274988  1.43769110  0.916939269
#>  [39,] -2.17927183 -0.511708924 -1.02988770 -2.397524012
#>  [40,]  1.65728442  1.001414511  0.45210221  0.481873041
#>  [41,] -0.73000431 -0.008149109 -0.63274935 -0.006960638
#>  [42,]  1.43383962  1.855431218 -0.37975507  0.266395227
#>  [43,]  0.22237624 -0.494655030  0.27335740 -0.715415914
#>  [44,]  0.11320698  0.554628152 -0.08770879  0.399203737
#>  [45,] -1.96078606 -0.746770496 -0.34843920 -0.537059436
#>  [46,]  1.65732661  0.492523610  1.06498691  0.568953018
#>  [47,] -0.12230917  1.976330180 -0.29837586 -0.354563836
#>  [48,]  2.30805679  1.621201271  1.62567206 -0.243682921
#>  [49,]  0.26207311  0.299292932  0.70705915  1.618666180
#>  [50,] -1.26096902 -0.189226276 -1.10923198 -0.579525409
#>  [51,] -0.22428128 -0.099576829  1.90898391  0.300757736
#>  [52,] -1.58219966  1.240504610 -0.54828468 -1.458138142
#>  [53,] -0.06898218 -0.851561803  0.98733869 -0.654196955
#>  [54,] -0.91785694 -2.059455620 -0.80995557  0.976057462
#>  [55,] -0.69268188  1.075116816 -0.49477880 -1.036894738
#>  [56,]  0.82664973  1.706154312 -0.13746534  0.562790498
#>  [57,]  0.39412376 -0.858248371  1.58314532 -0.060336914
#>  [58,]  0.20496676  1.307091232 -0.15488688  1.287361420
#>  [59,] -0.70452763 -1.412204547 -1.27621576  0.482468990
#>  [60,]  1.28034516  1.424835257  1.31037652  0.540759339
#>  [61,] -0.54903440 -0.096085369 -2.77664338 -0.038573992
#>  [62,]  1.01348019  0.587852276  0.97744039  2.002751117
#>  [63,] -0.53860656 -0.438562961  0.12406926  0.483548490
#>  [64,] -0.74710695  0.344322038  0.82381172 -0.413114576
#>  [65,]  1.20850654 -0.282182773  1.10933942  0.859734848
#>  [66,] -0.73462999 -0.014095648  0.13430260  1.493744037
#>  [67,] -1.09631934 -0.284503448 -1.40886330 -0.451078092
#>  [68,] -0.41907225 -0.555208253  0.34587423 -1.636743244
#>  [69,]  0.99716360 -0.832669091  0.78023288  0.870609527
#>  [70,] -1.25496661 -1.093741970 -0.65055041 -0.798265627
#>  [71,]  0.59029608 -0.008382627  0.02382613  0.014746397
#>  [72,] -0.68303509 -0.121436937 -1.35585713  0.400082882
#>  [73,]  1.13714841  0.211553305  0.33877348  0.975621984
#>  [74,]  1.68275648  0.771890476 -0.81090325  1.599400775
#>  [75,]  0.80155507  1.294783153  0.18925870  0.192576922
#>  [76,]  0.95950206 -0.468140379  1.19791708  0.682116870
#>  [77,]  2.15052010  0.582690913  0.42460085  1.347089271
#>  [78,]  0.47682545  1.257749400 -0.20895921  0.193700267
#>  [79,] -0.16077879  0.172850025 -0.11865449  0.094584826
#>  [80,] -1.30033662 -1.673403087 -1.06835843 -1.521051483
#>  [81,] -0.14587253  1.554860195 -0.52617382 -0.674018399
#>  [82,] -0.61142832 -1.487424407 -0.77574774 -0.217822609
#>  [83,] -1.09759888 -0.104281132 -0.07369154 -2.461489872
#>  [84,]  1.08610672 -0.425208762 -0.41836566 -0.151696634
#>  [85,] -1.41614890  0.402981150 -0.87012761 -0.224157952
#>  [86,]  2.19710275 -0.227515861 -0.71815704  1.712151035
#>  [87,] -1.43910847 -1.140742644 -0.59663141 -1.576966952
#>  [88,]  1.08259548  1.187626890  0.56309043  0.675579696
#>  [89,]  0.39032376  1.627367795 -0.22829086 -0.611846808
#>  [90,]  1.81542238  1.144058350  0.58247762  1.793697023
#>  [91,]  0.41204941  0.307187063  0.09097244  0.616552678
#>  [92,]  0.93590396  1.184706947  0.35791742  0.390555196
#>  [93,]  0.86903434  0.179128975  0.20328748  1.059967801
#>  [94,]  1.57042425 -0.523976446  1.15233996  1.512674257
#>  [95,] -0.13991038  1.506191559 -0.68122314  0.078506021
#>  [96,]  0.61167852 -0.011913569  0.66335850 -0.743613896
#>  [97,] -0.20413904 -0.372946887 -0.82721045  0.122183440
#>  [98,] -0.15434812 -1.077843637  0.29396946 -0.669795031
#>  [99,] -0.06916075  0.065514431 -0.29619355  0.220487609
#> [100,]  0.23640420 -0.686271587  0.92325616 -0.473269845

# }