Create multivariate data by permutation

corvectors() is a function to obtain a multivariate dataset by specifying the relation between those specified variables.

corvectors(
  data,
  corm,
  tol = 0.005,
  conv = 10000,
  cores = 2,
  splitsize = 1000,
  verbose = FALSE,
  seed
)

Arguments

data
a data matrix containing the data

corm
A value containing the desired correlation or a vector or data matrix containing the desired correlations

tol
A single value or a vector of tolerances with length ncol(data) - 1. The default is 0.005

conv
The maximum iterations allowed. Defaults to 1000.

cores
The number of cores to be used for parallel computing

splitsize
The size to use for splitting the data

verbose
Logical statement. Default is FALSE

seed
An optional seed to set

Value

corvectors() returns a matrix given the specified multivariate relation.

Details

This is liberally copy-pasted from van Kooten and Vink's wonderful-but-no-longer-supported correlate package. They call it correlate() in their package, but I opt for corvectors() here.

Author

Pascal van Kooten and Gerko Vink

Examples

# \donttest{
# bivariate example, start with zero correlation
corvectors(replicate(2, rnorm(100)), .5)

#>               [,1]        [,2]
#>   [1,]  0.20401676  1.27335615
#>   [2,]  0.94812875 -0.01523648
#>   [3,] -0.09757365  0.58415928
#>   [4,] -0.06665350 -0.19232924
#>   [5,] -1.29654378  0.17915543
#>   [6,]  1.14885697  0.63420154
#>   [7,]  0.21533313 -1.12558527
#>   [8,] -1.80303954  0.50315772
#>   [9,]  0.64325952  0.31243750
#>  [10,]  1.84601226 -0.58454153
#>  [11,]  0.97439470  2.74885198
#>  [12,] -1.81887642 -1.83357608
#>  [13,] -1.62721113 -1.16226101
#>  [14,]  0.39607245 -0.19944165
#>  [15,] -1.37798670  1.40618777
#>  [16,] -0.83372758 -2.09110033
#>  [17,] -0.22191503 -0.52564039
#>  [18,]  1.24298865  0.40015727
#>  [19,] -0.50127757  0.52698723
#>  [20,] -0.94238578  1.67444832
#>  [21,]  0.36841078 -0.05673319
#>  [22,] -0.27931928 -0.30465688
#>  [23,]  0.05349542 -0.64475668
#>  [24,]  0.69005635  0.88528647
#>  [25,]  0.05717782 -0.30940725
#>  [26,]  1.53677480  0.62760451
#>  [27,] -0.80488027 -1.77461569
#>  [28,]  1.47606428  1.66760165
#>  [29,]  0.71258675 -0.50103961
#>  [30,] -0.21208884 -0.24417342
#>  [31,]  0.75867769  1.55156748
#>  [32,] -0.46277729  0.67368064
#>  [33,]  0.83242230  1.34044840
#>  [34,]  1.08873598 -0.07750108
#>  [35,]  0.58584400  0.12069883
#>  [36,]  1.12966650  2.84297711
#>  [37,] -0.87586766  0.23723810
#>  [38,]  0.86323943 -0.46607721
#>  [39,]  0.18068492 -0.59279926
#>  [40,] -0.27178909 -0.57851390
#>  [41,] -0.79544564 -0.85451732
#>  [42,] -1.64928906 -1.99372143
#>  [43,]  1.26771349  2.82811446
#>  [44,]  0.12733224  2.44064189
#>  [45,]  0.73428443 -0.39862179
#>  [46,]  0.40128442  0.66873200
#>  [47,]  2.30398889  1.17282506
#>  [48,] -1.01059002  0.20946459
#>  [49,]  0.48464660 -0.05013828
#>  [50,] -0.86264027 -0.48212424
#>  [51,]  0.15931532  0.55382939
#>  [52,] -1.23713996  0.78295517
#>  [53,]  0.63667947 -0.03211670
#>  [54,]  0.40534701  0.23607535
#>  [55,]  1.15357665  0.05235419
#>  [56,]  0.57964236  0.73471324
#>  [57,]  0.87752485 -0.22757969
#>  [58,]  0.85023624 -1.21288296
#>  [59,]  0.07063484  0.90039651
#>  [60,] -0.05342496  0.15639488
#>  [61,]  0.95223303  1.78104065
#>  [62,]  0.15710255  0.68882635
#>  [63,] -1.32556083 -0.23104140
#>  [64,] -0.57787399 -1.15436535
#>  [65,]  0.59378781  1.33375112
#>  [66,] -1.43988413 -1.34523643
#>  [67,] -0.60372181 -0.46623541
#>  [68,]  0.82465469  1.37146581
#>  [69,] -1.47331528 -1.65526740
#>  [70,]  0.14214254  0.18838184
#>  [71,] -0.66795544 -1.74999650
#>  [72,]  0.09183975 -0.54074261
#>  [73,]  0.96041726  0.70001274
#>  [74,] -0.59425060  0.15758900
#>  [75,]  0.11678369 -0.65190423
#>  [76,]  0.69523628  0.20483938
#>  [77,] -1.62339383 -0.36960776
#>  [78,]  0.55738252 -0.76755218
#>  [79,] -1.19123927 -0.95706074
#>  [80,]  0.74541620  0.43600571
#>  [81,]  0.42370435  0.09252502
#>  [82,]  2.21357740  1.47291735
#>  [83,]  0.02838126  0.58007782
#>  [84,] -1.01238384 -2.27522072
#>  [85,] -0.95543146 -0.33489310
#>  [86,] -0.72324656 -1.25214027
#>  [87,] -0.02985419 -0.59044355
#>  [88,] -1.34454748  0.01691176
#>  [89,] -1.16394046 -2.68897992
#>  [90,] -2.03005826 -0.45845865
#>  [91,]  0.87577954  0.22893182
#>  [92,]  0.27503126  1.09260573
#>  [93,] -0.18013489  0.74526138
#>  [94,]  1.81787943 -0.38345966
#>  [95,]  0.57382730  0.58784717
#>  [96,] -0.38173896 -0.44799849
#>  [97,] -0.88329743 -0.49884358
#>  [98,] -1.11557458 -2.42406367
#>  [99,] -1.83473442 -1.27624821
#> [100,]  1.40223138  0.36985875

# multivariate example

corvectors(replicate(4, rnorm(100)), c(.5, .6, .7))

#>               [,1]         [,2]        [,3]        [,4]
#>   [1,]  1.37882101  1.918093056  2.97760117  0.93517642
#>   [2,] -0.50217242 -2.082571370  1.07082325 -1.14939131
#>   [3,]  0.65950956  0.481435621  0.82554472  0.47646938
#>   [4,]  0.56627327  0.246829476  1.96704947  0.95720153
#>   [5,] -0.03289600  0.455197052  2.58682017  0.21109531
#>   [6,] -0.39883558 -0.085007196 -0.88071836 -0.19626960
#>   [7,] -0.40906915  0.718251079 -0.65934291  0.21101355
#>   [8,] -0.31353984 -1.114362697 -0.57921140 -0.30400299
#>   [9,]  0.49649614  1.509388105 -0.13023899  0.27327238
#>  [10,]  1.36924097  1.570279666  1.15321544 -0.17204185
#>  [11,] -1.18715407 -0.719905583 -0.76922217 -1.31204722
#>  [12,]  0.21192888  0.461003577 -1.29292709  0.82821767
#>  [13,] -0.21375288  1.352112926  0.09638773  0.74093300
#>  [14,]  0.97230166  1.286609581  1.44651227  1.25207413
#>  [15,] -1.07266019 -1.086431609 -0.62821770 -1.89247835
#>  [16,] -0.25780391 -0.727642652 -0.53852740  0.83953795
#>  [17,]  2.15472480  1.114680096  1.21181140  0.45135036
#>  [18,]  0.33953929  1.088660909  0.75157968 -0.04983286
#>  [19,] -0.69526372  0.673820782 -0.05171429 -0.58094554
#>  [20,]  0.14820592 -1.145390500 -0.47420171  0.21410239
#>  [21,] -1.18518027 -0.141279367 -1.66879778  0.11839306
#>  [22,] -0.59177966  0.656818074 -0.85872635 -1.17817342
#>  [23,] -0.54684545 -0.506979043  0.54937551 -1.02372318
#>  [24,] -0.69752677 -1.067964199 -1.24061523 -1.76244517
#>  [25,] -0.48179695  1.342194521  2.05577233 -0.90299244
#>  [26,]  0.06088806 -0.251614174 -1.16863770  0.39636283
#>  [27,]  1.97875227  1.520000705  0.96330162  1.60954250
#>  [28,]  1.37722900  2.330663529 -0.04397552  0.36620092
#>  [29,] -0.03334329 -2.037947106 -1.13654005 -0.22695416
#>  [30,] -0.01701577  0.382262127  0.20078392 -1.34668999
#>  [31,] -0.67688177 -1.452496715 -1.15748758 -1.71417842
#>  [32,]  0.36926053  0.475933255  0.07612191 -0.57517564
#>  [33,] -0.78294302  0.177279658 -0.16509609 -0.24019534
#>  [34,]  1.18548536 -0.230983296  1.59234552  2.48296426
#>  [35,]  0.54960469  1.139778817  0.50139469  0.40938441
#>  [36,]  0.75394782 -0.150518296 -0.30292451  1.39796133
#>  [37,]  0.61168506  1.232745732  0.73373866 -0.69745837
#>  [38,] -0.19749526 -1.445512796  0.72657171  0.15118450
#>  [39,] -0.38589920 -0.807893005 -1.73305014 -1.04749353
#>  [40,] -0.31741281 -0.809420079 -0.19336159  0.49147644
#>  [41,] -0.08375949  0.600738853 -0.07097449 -0.90442465
#>  [42,] -0.50449867 -2.079965319 -1.26142520 -0.55443134
#>  [43,]  0.70681160  0.753170488  0.16025346  0.78722047
#>  [44,]  0.20171080  0.855714633 -0.52745635  0.65918277
#>  [45,] -0.07991034 -0.557576972  1.47580282 -0.14458542
#>  [46,]  0.74453663 -0.005247644  1.17928123 -0.38680661
#>  [47,] -1.03272075  0.364924190 -1.26584476 -1.56643946
#>  [48,] -0.56740368  0.654086854  0.46252005 -0.87511657
#>  [49,]  1.12383352  0.968243540  0.96601248  0.10890525
#>  [50,] -1.02302274 -0.988451685 -0.74521193 -1.00975698
#>  [51,]  0.71349369  0.273721421  0.48540411  0.46941911
#>  [52,] -0.49025726 -0.062588582 -1.26375141 -0.15636378
#>  [53,] -0.99660266 -0.485440771 -0.48338887 -1.44716011
#>  [54,] -0.50580244 -0.527120072 -0.60011522 -0.92583493
#>  [55,]  0.82044275  0.041088330  2.61489098 -0.20893628
#>  [56,]  0.94391839 -0.447311502  0.54828642  0.72441262
#>  [57,] -0.54328686  0.664754985 -0.26620594 -1.22937204
#>  [58,]  0.96813616  2.083686203 -0.14071840 -0.22062137
#>  [59,]  0.83449171 -0.453304926  1.50174756  0.88014578
#>  [60,]  0.06711674  2.173656467 -1.60823688  0.02970100
#>  [61,] -0.74670296  1.465756871 -2.08485833 -1.34187456
#>  [62,]  0.26965846  0.685780550 -0.39441161  0.68445129
#>  [63,]  0.51796654  0.509023493  0.64459993  1.16015775
#>  [64,]  1.19515619  1.390873967  1.82054482  1.12633307
#>  [65,]  2.20132989  1.850467238  0.78417907  1.15729264
#>  [66,] -0.34520574  0.486249963 -1.14958610  0.99810557
#>  [67,]  0.61289503  0.795109485  0.08941668 -0.61481366
#>  [68,]  0.71109618  0.403718812  2.36444490  0.65439086
#>  [69,]  0.23947827  1.147459407  0.40746097 -0.69694031
#>  [70,]  0.28001623 -0.332960529  0.07294633  0.68919149
#>  [71,] -0.59694216 -0.163360583 -0.66549186 -0.60428425
#>  [72,]  0.22003206 -0.359506597  0.61943392  0.18547399
#>  [73,]  1.17547755 -0.151765725 -0.46564832  0.98316120
#>  [74,]  0.33173722  0.325363858 -0.18834937  0.86662628
#>  [75,]  0.73308854  0.294833116  0.25806322 -0.44572936
#>  [76,]  0.23501039 -0.467657881  0.94591070  1.36975137
#>  [77,]  0.23369942 -1.719648764  0.84053550  0.23627247
#>  [78,] -0.19095888 -0.189518510 -0.26326803 -0.07357670
#>  [79,]  1.57666012  0.796690790  1.63595950  2.23782667
#>  [80,]  0.23031097  0.221575566  0.05152193 -0.92640662
#>  [81,] -0.74823734 -0.529749271 -0.86470282 -0.43627047
#>  [82,] -0.43907490 -0.340232223 -1.21086823 -0.91136757
#>  [83,] -0.03559946  0.024353327 -0.23682587  0.49855271
#>  [84,]  1.50345406  1.434723958  1.37922792  1.84971340
#>  [85,]  2.15378704 -0.431517083  1.52888669  1.22703298
#>  [86,]  0.42882995 -0.213419280  0.45703709  0.36315315
#>  [87,] -0.33458634 -0.177188908 -0.60892336 -0.41308673
#>  [88,]  0.07736876 -0.436188463  0.45915435 -0.52948357
#>  [89,] -0.65930755 -1.251221302 -0.86832782 -2.17175636
#>  [90,] -0.79719092  0.291841451 -0.70794318 -1.64652422
#>  [91,]  0.09088352  0.756634660 -0.91968276 -0.44661763
#>  [92,] -0.43166855  0.443781502 -0.36936251 -0.04533354
#>  [93,] -0.12386386  0.284174273  1.33072304  1.10022031
#>  [94,]  0.23200831  0.961864384 -0.10375857 -0.82314584
#>  [95,] -0.20791026  0.068083587 -1.31647501 -1.61625946
#>  [96,] -0.41065681 -1.904507096  1.32845567 -0.18554990
#>  [97,]  0.84386474  1.217297291  0.49339102  0.52226455
#>  [98,] -1.38198298 -1.440319800 -0.85463740 -2.07465859
#>  [99,]  1.10050745  2.122255835  3.36994084  1.00305216
#> [100,] -0.35390369  0.414307937  0.35960313 -0.46853632

# }