Create multivariate data by permutation

corvectors() is a function to obtain a multivariate dataset by specifying the relation between those specified variables.

corvectors(
  data,
  corm,
  tol = 0.005,
  conv = 10000,
  cores = 2,
  splitsize = 1000,
  verbose = FALSE,
  seed
)

Arguments

data
a data matrix containing the data

corm
A value containing the desired correlation or a vector or data matrix containing the desired correlations

tol
A single value or a vector of tolerances with length ncol(data) - 1. The default is 0.005

conv
The maximum iterations allowed. Defaults to 1000.

cores
The number of cores to be used for parallel computing

splitsize
The size to use for splitting the data

verbose
Logical statement. Default is FALSE

seed
An optional seed to set

Value

corvectors() returns a matrix given the specified multivariate relation.

Details

This is liberally copy-pasted from van Kooten and Vink's wonderful-but-no-longer-supported correlate package. They call it correlate() in their package, but I opt for corvectors() here.

Author

Pascal van Kooten and Gerko Vink

Examples

# \donttest{
# bivariate example, start with zero correlation
corvectors(replicate(2, rnorm(100)), .5)

#>               [,1]        [,2]
#>   [1,]  0.71258675 -0.50103961
#>   [2,]  0.15710255  0.68882635
#>   [3,] -1.19123927 -0.95706074
#>   [4,] -1.37798670 -0.24417342
#>   [5,]  0.86323943  0.15639488
#>   [6,] -0.57787399  1.47291735
#>   [7,] -2.03005826 -1.25214027
#>   [8,] -0.27178909  1.17282506
#>   [9,]  0.58584400  0.12069883
#>  [10,]  0.87577954  0.22893182
#>  [11,]  0.63667947 -0.03211670
#>  [12,]  0.02838126  0.58007782
#>  [13,]  0.87752485  2.74885198
#>  [14,] -0.59425060  0.55382939
#>  [15,]  1.24298865  0.40015727
#>  [16,]  0.09183975 -0.54074261
#>  [17,]  1.08873598  0.58784717
#>  [18,]  0.39607245  0.23607535
#>  [19,] -0.38173896  0.78295517
#>  [20,]  0.59378781  0.01691176
#>  [21,] -0.27931928 -0.30465688
#>  [22,]  0.05349542 -0.64475668
#>  [23,]  0.11678369 -0.65190423
#>  [24,]  0.64325952  0.31243750
#>  [25,]  0.97439470  0.52698723
#>  [26,]  0.69523628 -0.58454153
#>  [27,] -0.88329743 -1.77461569
#>  [28,]  0.55738252 -0.59044355
#>  [29,]  0.57964236  1.37146581
#>  [30,] -0.18013489 -1.83357608
#>  [31,]  1.14885697  0.70001274
#>  [32,] -0.46277729  0.67368064
#>  [33,]  0.75867769  0.15758900
#>  [34,] -1.23713996 -0.30940725
#>  [35,] -0.95543146  0.05235419
#>  [36,]  0.94812875 -0.01523648
#>  [37,] -0.22191503 -0.52564039
#>  [38,]  0.96041726 -0.22757969
#>  [39,] -0.72324656 -0.05673319
#>  [40,]  0.21533313 -0.85451732
#>  [41,]  0.12733224  2.44064189
#>  [42,] -1.47331528 -1.65526740
#>  [43,]  0.48464660  0.66873200
#>  [44,] -1.64928906 -2.68897992
#>  [45,]  0.15931532  0.20483938
#>  [46,] -1.32556083  0.50315772
#>  [47,] -0.02985419 -1.99372143
#>  [48,] -1.01059002  0.20946459
#>  [49,]  1.53677480  1.78104065
#>  [50,]  0.57382730 -0.07750108
#>  [51,]  1.15357665  1.55156748
#>  [52,] -0.09757365  0.58415928
#>  [53,]  1.12966650  2.84297711
#>  [54,]  1.81787943 -0.38345966
#>  [55,] -0.86264027 -1.21288296
#>  [56,]  1.84601226  0.90039651
#>  [57,]  0.85023624 -0.46607721
#>  [58,]  0.82465469  0.73471324
#>  [59,] -1.34454748 -2.09110033
#>  [60,]  0.20401676  1.27335615
#>  [61,]  2.30398889  1.33375112
#>  [62,] -1.29654378  0.17915543
#>  [63,] -1.43988413  0.63420154
#>  [64,]  0.95223303 -0.44799849
#>  [65,]  0.27503126  0.62760451
#>  [66,]  0.18068492 -2.27522072
#>  [67,]  0.40128442 -0.57851390
#>  [68,]  1.47606428  1.66760165
#>  [69,] -1.16394046 -0.76755218
#>  [70,] -0.80488027 -0.19232924
#>  [71,]  0.40534701 -0.19944165
#>  [72,] -1.81887642  0.36985875
#>  [73,] -1.80303954 -1.34523643
#>  [74,]  0.07063484  0.23723810
#>  [75,] -0.50127757 -1.15436535
#>  [76,]  0.05717782  1.67444832
#>  [77,] -1.62721113 -1.16226101
#>  [78,] -1.01238384 -0.05013828
#>  [79,] -0.05342496 -0.39862179
#>  [80,]  1.26771349  1.09260573
#>  [81,] -0.83372758 -1.27624821
#>  [82,] -0.94238578 -0.23104140
#>  [83,]  2.21357740  2.82811446
#>  [84,]  0.14214254  0.18838184
#>  [85,] -0.66795544 -1.74999650
#>  [86,] -0.60372181 -0.46623541
#>  [87,]  1.40223138 -0.59279926
#>  [88,] -0.21208884  1.40618777
#>  [89,] -1.83473442 -0.45845865
#>  [90,] -0.87586766 -2.42406367
#>  [91,]  0.42370435  0.74526138
#>  [92,] -1.11557458 -0.33489310
#>  [93,]  0.83242230  1.34044840
#>  [94,] -1.62339383 -0.36960776
#>  [95,]  0.73428443 -0.48212424
#>  [96,] -0.79544564 -1.12558527
#>  [97,]  0.36841078  0.09252502
#>  [98,]  0.74541620  0.43600571
#>  [99,]  0.69005635  0.88528647
#> [100,] -0.06665350 -0.49884358

# multivariate example

corvectors(replicate(4, rnorm(100)), c(.5, .6, .7))

#>               [,1]          [,2]        [,3]        [,4]
#>   [1,] -1.37138600 -1.5581079182 -1.43256541 -0.98207803
#>   [2,] -1.40657270 -0.1240873481 -0.36733317 -1.80571328
#>   [3,] -1.57479453 -0.0577985765  0.07450573 -2.01016163
#>   [4,] -1.35949430 -0.8655774730  0.21681324 -0.43691114
#>   [5,]  2.63702198  0.6118348662  1.91797360  2.07097779
#>   [6,] -1.01659737  0.1487794597 -0.45299585 -1.32735909
#>   [7,]  2.46158496  0.2415867947  1.92581982  0.70515180
#>   [8,] -0.38778273 -1.6254394597 -2.66390252 -2.29665551
#>   [9,] -1.25310923 -0.5456140286  0.79507291 -0.87818160
#>  [10,] -0.53761370 -0.9538672719  0.59697047 -0.23630099
#>  [11,]  1.05219768  0.7627511726  1.12770630  0.99824803
#>  [12,] -0.67671648  0.2181233494 -0.69973530 -1.45853279
#>  [13,] -2.11029480 -1.4599688995 -1.08754992 -0.82843930
#>  [14,] -0.68195534 -0.2356272164  0.44691751  0.29350409
#>  [15,]  1.32307524  1.6973770978  1.66686308  0.57875152
#>  [16,]  0.77918774  1.0238233430  0.74854692  0.68324079
#>  [17,]  0.09019305  0.9629150348  1.00198235  0.96570756
#>  [18,]  0.05326493  0.9403737083  1.08974505 -0.92832504
#>  [19,] -0.74744442 -0.6728544658  0.64735028 -1.62798611
#>  [20,] -0.95745550 -2.3493743763 -0.20031125 -1.38729182
#>  [21,]  1.94609717  0.2859219307  2.01164066  1.10322609
#>  [22,]  0.39441911  0.1878873330  0.02153617 -0.88216098
#>  [23,]  0.54979738 -0.6418591737  0.81224776 -0.45775005
#>  [24,] -0.72877434 -1.3265596845 -0.94084160 -2.48914967
#>  [25,]  0.52275573 -0.8970311528 -0.99359893  0.63870157
#>  [26,] -0.84282368  0.8286152849  0.37074333 -0.87521913
#>  [27,] -1.19789594 -3.0894352360 -0.61669720 -1.17626938
#>  [28,] -0.61871197 -0.0794445979 -0.26542525 -0.01891234
#>  [29,]  0.97970469  0.1333107197  1.34212122  0.45237890
#>  [30,]  0.20485154  0.3008584329  0.60726051  2.15890479
#>  [31,]  0.87823360 -0.9096406402  1.30167123 -0.41365171
#>  [32,]  0.59881664 -1.1882322835  0.79158308  1.07852706
#>  [33,] -0.78157737 -1.1163058225  0.32591825 -1.13833880
#>  [34,]  1.18029960  0.4822020837 -0.53651222  1.57352595
#>  [35,] -0.49125533  0.6116691758  0.03836601 -1.59645852
#>  [36,] -0.41235823  0.8152939461 -0.76625637 -0.19066186
#>  [37,] -1.00253494 -0.5943657383  0.56191433  0.47238906
#>  [38,]  0.30443159 -0.2506029746  0.65216734  1.14721768
#>  [39,] -0.60652492  0.5095251013 -0.61154601 -0.35976264
#>  [40,]  0.96020954 -0.0261211558  0.70028468  0.85816471
#>  [41,]  0.53259254 -0.3274316434  1.72208083  1.94984676
#>  [42,] -0.21921244 -0.4592829994  1.01878641  0.29806552
#>  [43,]  0.54088533  1.5018116681  1.27810397  0.67310723
#>  [44,] -1.08123631 -0.5862873480 -1.15882355 -1.05256424
#>  [45,] -0.89015567 -0.8455728641 -0.32525890 -0.16013429
#>  [46,]  0.33233584 -0.9340483852  0.31225666 -0.42697435
#>  [47,] -0.36188198  0.5362566452 -0.62901425  0.77179948
#>  [48,] -0.39847044 -0.7149206208 -0.68910646 -0.27653968
#>  [49,]  1.14507853  0.3177095208 -0.15387397  1.84900220
#>  [50,] -1.35655088 -0.8308306171 -1.00970655 -1.54419436
#>  [51,]  0.18513027  0.7755412508  0.17746061  0.54816134
#>  [52,]  0.80615393 -0.1652384395 -0.82768915  1.20428630
#>  [53,] -1.00461589  0.0375011369 -0.55662979 -0.99237410
#>  [54,] -1.10455689  0.3848072394 -0.20380532 -0.97043332
#>  [55,]  1.32272035  0.4061677289  1.65117237 -0.01428331
#>  [56,]  1.69778009  0.8821007808  0.55356382  1.49694733
#>  [57,]  1.05753055  1.0509399192  1.17214317  2.18331826
#>  [58,]  0.32668682 -0.4669223683  0.60673796  0.21862586
#>  [59,]  0.35930351 -1.5131367373  0.51496278  0.46363897
#>  [60,] -0.27329385 -0.8079519619 -1.30526405  0.03753717
#>  [61,] -1.62894187 -2.0437915398 -0.20298343 -2.10939396
#>  [62,] -1.04351781 -1.2418506418  0.30612905 -0.48381601
#>  [63,] -0.84145897  0.1817608665 -0.99550872 -0.20268002
#>  [64,] -0.29747219 -0.3566609226 -0.64983275  0.12542505
#>  [65,] -1.20972138 -0.4691859180 -0.94185589 -0.37820475
#>  [66,]  0.47004143  0.0003354162  0.58400667  0.81393846
#>  [67,] -0.20731681  0.3750084287 -0.05690401 -0.11183763
#>  [68,]  0.52575112  1.9643038332  0.31079782  1.12767045
#>  [69,] -1.22299757 -0.2104967419 -2.36862038 -1.37314045
#>  [70,] -1.01701460 -0.7845585070 -2.59893783 -1.17548018
#>  [71,] -0.75049498  0.3799635345 -0.07374903  0.10985027
#>  [72,] -0.14298005 -0.2126370201  0.49176899  0.88665229
#>  [73,]  2.40033631  0.6698955476  1.04795276  1.72527111
#>  [74,] -0.10930596  0.1815201563  0.16177237  0.29088653
#>  [75,]  0.18098977  1.3289726614  1.51657069 -1.03702017
#>  [76,]  0.69944094  2.4998958986 -0.08117769  0.56476553
#>  [77,] -0.37456531  0.4335510968 -1.05970756  0.67342119
#>  [78,] -1.59015583 -1.5686406094 -1.63546723 -0.55705465
#>  [79,]  1.63518955  1.1742713573  1.31831936  0.23434464
#>  [80,] -0.79196149 -0.9672961182 -1.42865261 -0.22548144
#>  [81,] -0.36606413 -0.1060391712  1.37696889 -1.61443021
#>  [82,]  0.07568805 -2.2255636442  0.13370272  1.18702740
#>  [83,]  0.92470161  2.0191359215  0.81244842  0.84174985
#>  [84,]  0.45335177  0.3312353363 -0.50328651 -0.34221742
#>  [85,]  0.02721972  0.8707731056  0.36017218  1.00792330
#>  [86,] -0.14907846  0.5835806337  2.24264440  0.40881311
#>  [87,]  0.44392701  0.1680478492 -0.30826867  0.28989508
#>  [88,] -0.61289399  0.2907359833  0.59454180  0.12945410
#>  [89,]  0.61047264  0.5803942882  0.53340028  0.92526960
#>  [90,]  0.86776719  0.9225958161  1.28645452  0.91155966
#>  [91,] -0.98647138  0.5773073256 -0.30199837 -0.54465000
#>  [92,] -0.40828894  0.2565472463  0.49903189 -0.48854019
#>  [93,] -0.21338344  0.5975467169  0.18170594 -0.48440975
#>  [94,] -0.56398978 -0.0910310274 -0.57663880 -1.26805826
#>  [95,]  1.65971118  1.6235174760  2.55104456  1.17885523
#>  [96,]  0.57937085  1.5119288453 -0.68608953  0.08587102
#>  [97,] -0.84408829  0.2325732736  0.78325213  0.71308733
#>  [98,] -0.41468476 -1.3967843352  0.73542438 -0.54737313
#>  [99,] -0.17256468  0.4494456385  0.37413219  0.36782452
#> [100,] -0.36742194 -1.1463802331  0.15500160  0.25759728

# }