Create multivariate data by permutation

corvectors() is a function to obtain a multivariate dataset by specifying the relation between those specified variables.

corvectors(
  data,
  corm,
  tol = 0.005,
  conv = 10000,
  cores = 2,
  splitsize = 1000,
  verbose = FALSE,
  seed
)

Arguments

data
a data matrix containing the data

corm
A value containing the desired correlation or a vector or data matrix containing the desired correlations

tol
A single value or a vector of tolerances with length ncol(data) - 1. The default is 0.005

conv
The maximum iterations allowed. Defaults to 1000.

cores
The number of cores to be used for parallel computing

splitsize
The size to use for splitting the data

verbose
Logical statement. Default is FALSE

seed
An optional seed to set

Value

corvectors() returns a matrix given the specified multivariate relation.

Details

This is liberally copy-pasted from van Kooten and Vink's wonderful-but-no-longer-supported correlate package. They call it correlate() in their package, but I opt for corvectors() here.

Author

Pascal van Kooten and Gerko Vink

Examples

# \donttest{
set.seed(8675309)
library(tibble)
# bivariate example, start with zero correlation
as_tibble(data.frame(corvectors(replicate(2, rnorm(100)), .5)))

#> # A tibble: 100 x 2
#>          X1     X2
#>       <dbl>  <dbl>
#>  1  0.722    0.303
#>  2  0.684   -0.958
#>  3 -1.31     0.990
#>  4  1.57     1.32 
#>  5  0.214   -0.605
#>  6 -0.415   -1.35 
#>  7 -0.652   -2.09 
#>  8 -0.151    0.711
#>  9 -0.00379 -1.70 
#> 10 -0.926   -1.03 
#> # … with 90 more rows

# multivariate example

as_tibble(data.frame(corvectors(replicate(4, rnorm(100)), c(.5, .6, .7))))

#> # A tibble: 100 x 4
#>         X1     X2      X3      X4
#>      <dbl>  <dbl>   <dbl>   <dbl>
#>  1 -0.105  -1.78   0.706  -0.530 
#>  2  0.0536 -0.525  0.425   0.0909
#>  3 -1.07   -1.73  -0.442  -1.37  
#>  4  0.726  -1.12   1.17    0.853 
#>  5 -0.192  -0.293  0.125  -0.889 
#>  6 -0.824  -1.61   0.0133 -1.67  
#>  7 -1.20   -0.765 -0.827  -0.0878
#>  8  0.918  -0.415  0.266   1.25  
#>  9 -0.332   0.811  0.231  -0.239 
#> 10  0.0995  0.482  1.65    0.542 
#> # … with 90 more rows

# }