Filter a Data Frame of Citations and Return the Entries as a Character

filter_refs() is a convenience function I wrote for filtering a data frame of citations returning the entries as a valid .bib entry (as a character vector). I wrote this for more easily passing on citations to the print_refs() function also included in this package.

filter_refs(bibdat, criteria, type = "bibtexkey")

Arguments

bibdat
a data frame of citations, like the one created by the bib2df package

criteria
criteria, specified as a character vector, by which to filter the data frame of citations

type
the particular type of citation entry on which to filter. Defaults to "bibtexkey" (which filters based on a column of unique citation keys). When type == "year", the function filters on a character vector of years.

Value

filter_refs() takes a data frame of citations, like the one created by the bib2df package, and returns a character vector (amounting to a valid .bib entry) of citations the user wants. This can then be easily passed to the print_refs() function also included in this package.

Details

filter_refs() assumes some familiarity with BibTeX, .bib entries, and depends on the bib2df package.

Examples

# Based on `stevepubs` configuration, filter on `BIBTEXKEY` where
# the citation key matches one of these.
filter_refs(stevepubs, c("miller2017etst", "miller2017etjc", "miller2013tdpi"))

#>  [1] "@Article{miller2013tdpi,"                                                                                        
#>  [2] "  Author = {Steven V. Miller},"                                                                                  
#>  [3] "  Journal = {Journal of Peace Research},"                                                                        
#>  [4] "  Number = {6},"                                                                                                 
#>  [5] "  Pages = {677--690},"                                                                                           
#>  [6] "  Title = {Territorial Disputes and the Politics of Individual Well-Being},"                                     
#>  [7] "  Volume = {50},"                                                                                                
#>  [8] "  Year = {2013}"                                                                                                 
#>  [9] "}"                                                                                                               
#> [10] ""                                                                                                                
#> [11] ""                                                                                                                
#> [12] "@Article{miller2017etst,"                                                                                        
#> [13] "  Author = {Steven V. Miller},"                                                                                  
#> [14] "  Journal = {Political Behavior},"                                                                               
#> [15] "  Number = {2},"                                                                                                 
#> [16] "  Pages = {457--478},"                                                                                           
#> [17] "  Title = {Economic Threats or Societal Turmoil? Understanding Preferences for Authoritarian Political Systems},"
#> [18] "  Volume = {39},"                                                                                                
#> [19] "  Year = {2017}"                                                                                                 
#> [20] "}"                                                                                                               
#> [21] ""                                                                                                                
#> [22] ""                                                                                                                
#> [23] "@Article{miller2017etjc,"                                                                                        
#> [24] "  Author = {Steven V. Miller},"                                                                                  
#> [25] "  Journal = {Political Research Quarterly},"                                                                     
#> [26] "  Number = {4},"                                                                                                 
#> [27] "  Pages = {790--802},"                                                                                           
#> [28] "  Title = {The Effect of Terrorism on Judicial Confidence},"                                                     
#> [29] "  Volume = {70},"                                                                                                
#> [30] "  Year = {2017}"                                                                                                 
#> [31] "}"                                                                                                               

# Based on `stevepubs` configuration, filter on `YEAR` where
# the publication year is 2017, 2018, 2019, 2020, or 2021.
filter_refs(stevepubs, c(2017:2021), type = "year")

#>  [1] "@Article{miller2017etst,"                                                                                        
#>  [2] "  Author = {Steven V. Miller},"                                                                                  
#>  [3] "  Journal = {Political Behavior},"                                                                               
#>  [4] "  Number = {2},"                                                                                                 
#>  [5] "  Pages = {457--478},"                                                                                           
#>  [6] "  Title = {Economic Threats or Societal Turmoil? Understanding Preferences for Authoritarian Political Systems},"
#>  [7] "  Volume = {39},"                                                                                                
#>  [8] "  Year = {2017}"                                                                                                 
#>  [9] "}"                                                                                                               
#> [10] ""                                                                                                                
#> [11] ""                                                                                                                
#> [12] "@Article{miller2017ieea,"                                                                                        
#> [13] "  Author = {Steven V. Miller},"                                                                                  
#> [14] "  Journal = {Conflict Management and Peace Science},"                                                            
#> [15] "  Number = {5},"                                                                                                 
#> [16] "  Pages = {526--545},"                                                                                           
#> [17] "  Title = {Individual-Level Expectations of Executive Authority under Territorial Threat},"                      
#> [18] "  Volume = {34},"                                                                                                
#> [19] "  Year = {2017}"                                                                                                 
#> [20] "}"                                                                                                               
#> [21] ""                                                                                                                
#> [22] ""                                                                                                                
#> [23] "@Article{miller2017etjc,"                                                                                        
#> [24] "  Author = {Steven V. Miller},"                                                                                  
#> [25] "  Journal = {Political Research Quarterly},"                                                                     
#> [26] "  Number = {4},"                                                                                                 
#> [27] "  Pages = {790--802},"                                                                                           
#> [28] "  Title = {The Effect of Terrorism on Judicial Confidence},"                                                     
#> [29] "  Volume = {70},"                                                                                                
#> [30] "  Year = {2017}"                                                                                                 
#> [31] "}"                                                                                                               
#> [32] ""                                                                                                                
#> [33] ""                                                                                                                
#> [34] "@Article{miller2018etttc,"                                                                                       
#> [35] "  Author = {Steven V. Miller},"                                                                                  
#> [36] "  Journal = {Peace Economics, Peace Science and Public Policy},"                                                 
#> [37] "  Number = {1},"                                                                                                 
#> [38] "  Title = {External Territorial Threats and Tolerance of Corruption: A Private/Government Distinction},"         
#> [39] "  Volume = {24},"                                                                                                
#> [40] "  Year = {2018},"                                                                                                
#> [41] "  Doi = {10.1515/peps-2017-0043}"                                                                                
#> [42] "}"                                                                                                               
#> [43] ""                                                                                                                
#> [44] ""                                                                                                                
#> [45] "@Article{miller2019wata,"                                                                                        
#> [46] "  Author = {Steven V. Miller},"                                                                                  
#> [47] "  Journal = {Social Science Quarterly},"                                                                         
#> [48] "  Number = {1},"                                                                                                 
#> [49] "  Pages = {272--288},"                                                                                           
#> [50] "  Title = {What Americans Think About Gun Control: Evidence from the General Social Survey, 1972-2016},"         
#> [51] "  Volume = {100},"                                                                                               
#> [52] "  Year = {2019}"                                                                                                 
#> [53] "}"                                                                                                               
#> [54] ""                                                                                                                
#> [55] ""                                                                                                                
#> [56] "@Article{gibleretal2020icm,"                                                                                     
#> [57] "  Author = {Douglas M. Gibler and Steven V. Miller and Erin K. Little},"                                         
#> [58] "  Journal = {International Studies Quarterly},"                                                                  
#> [59] "  Number = {2},"                                                                                                 
#> [60] "  Pages = {476--479},"                                                                                           
#> [61] "  Title = {The Importance of Correct Measurement},"                                                              
#> [62] "  Volume = {64},"                                                                                                
#> [63] "  Year = {2020}"                                                                                                 
#> [64] "}"                                                                                                               
#> [65] ""                                                                                                                
#> [66] ""                                                                                                                
#> [67] "@Incollection{milleretal2020gtc,"                                                                                
#> [68] "  Author = {Steven V. Miller and Jaroslav Tir and John A. Vasquez},"                                             
#> [69] "  Booktitle = {Oxford Research Encyclopedia of International Studies},"                                          
#> [70] "  Publisher = {Oxford University Press},"                                                                        
#> [71] "  Title = {Geography, Territory, and Conflict},"                                                                 
#> [72] "  Year = {2020},"                                                                                                
#> [73] "  Doi = {10.1093/acrefore/9780190846626.013.320}"                                                                
#> [74] "}"                                                                                                               
#> [75] ""                                                                                                                
#> [76] ""                                                                                                                
#> [77] "@Article{millerdavis2020ewsp,"                                                                                   
#> [78] "  Author = {Steven V. Miller and Nicholas T. Davis},"                                                            
#> [79] "  Journal = {Journal of Race, Ethnicity, and Politics},"                                                         
#> [80] "  Number = {2},"                                                                                                 
#> [81] "  Pages = {334--351},"                                                                                           
#> [82] "  Title = {The Effect of White Social Prejudice on Support for American Democracy},"                             
#> [83] "  Volume = {6},"                                                                                                 
#> [84] "  Year = {2021}"                                                                                                 
#> [85] "}"                                                                                                               
#> [86] ""                                                                                                                
#> [87] ""                                                                                                                
#> [88] "@Article{curtismiller2021snp,"                                                                                   
#> [89] "  Author = {K. Amber Curtis and Steven V. Miller},"                                                              
#> [90] "  Journal = {European Union Politics},"                                                                          
#> [91] "  Number = {2},"                                                                                                 
#> [92] "  Pages = {202--26},"                                                                                            
#> [93] "  Title = {A (Supra)Nationalist Personality? The Big Five’s Effects on Political-Territorial Identification},"   
#> [94] "  Volume = {22},"                                                                                                
#> [95] "  Year = {2021}"                                                                                                 
#> [96] "}"